% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adv_functions.R
\name{adv_managers_current_period_summary}
\alias{adv_managers_current_period_summary}
\title{ADV managers most recent summary data}
\usage{
adv_managers_current_period_summary(
  select_names = c("dateDataADV", "isExempt", "idRegionSEC", "idCRD", "idSEC",
    "typeRegulationSEC", "nameEntityManager", "nameEntityManagerLegal",
    "addressOfficePrimary", "addressStreet1OfficePrimary", "addressStreet2OfficePrimary",
    "cityOfficePrimary", "stateOfficePrimary", "countryOfficePrimary",
    "zipOfficePrimary", "phoneOfficePrimary", "statusSEC", "dateStatusSEC",
    "dateADVLatest", "urlManager", "isForeignRegisteredEntity",
    "stateDateJurisdictionNotice", "idCIK", "hasAUMGreater1B", "idLEI",
    "hasAUMGreater100M",      "typeEntity", "countryEntityOrganized",
    "countEmployeesTotal", "countEmployeesInvestmentAdvisory", "amountAUMTotal",
    "amountAUMDiscretionary", "amountAUMNonDiscretionary", "countAccountsDiscretionary",
    "countAccountsNonDiscretionary", "countAccountsTotal",
    "isManagerSecuritiesPortfolio", "hasFeeAUM", "hasFeeHourlyCharge",
    "hasFeeSubscription", "hasFeeFixed", "hasFeeCommission", "hasFeePerformance",
    "hasFeeOther", "typeFeeOther", "isBrokerDealer", "isBrokerDealerRepresentative",
    "isCommodityPoolOperator",      "isFuturesMerchant", "isRealEstateBrokerDealerAgent",
    "isInsuranceBrokerAgent", "isBank", "isTrustCompany", "isRegisteredMunicipalAdviser",
    "isRegisteredSecuritySwapDealer", "isRegistredSecuritySwapParticipant",
    "isAccountingFirm", "isLawFirm", "isOtherFinancialProductSalesperson",
    "typeOtherFinancialProductSalesperson", "countEmployeesBrokerDealer",
    "countEmployeesStateRegisteredInvestmentAdviser",
    "countEmployeesStateRegisteredInvestmentAdviserMultipleEntities",
    "countEmployeesLicensedInsuranceAgents",      "countEmployeesSolicitAdvisoryClients",
    "hasFelonyPleaConviction", "hasFelonyCharge", "hasMisdemeanorPleaConviction"),
  return_message = TRUE
)
}
\arguments{
\item{return_message}{return a message after parsing data}

\item{file_directory}{directory you want to save your data into, if none specified a temporary file will be created}

\item{folder_name}{older you want the data to be downloaded into}

\item{remove_files}{remove the files from the folders}

\item{empty_trash}{empty the trash after being read into R}
}
\value{
a data frame
}
\description{
This function returns abbreviated ADV data for
all filing managers for the most recent period.
}
\details{
For multiple periods and all information see \code{\link{adv_managers_periods_summaries}}
}
\examples{
\dontrun{
adv_managers_current_period_summary(select_names = c("dateDataADV", "isExempt", "idRegionSEC", "idCRD", "idSEC", "typeRegulationSEC", "nameEntityManager", "nameEntityManagerLegal", 'addressOfficePrimary', "addressStreet1OfficePrimary", "addressStreet2OfficePrimary", "cityOfficePrimary", "stateOfficePrimary", "countryOfficePrimary", "zipOfficePrimary", "phoneOfficePrimary", "statusSEC", "dateStatusSEC", "dateADVLatest", "urlManager", "isForeignRegisteredEntity", "stateDateJurisdictionNotice", "idCIK", "hasAUMGreater1B", "idLEI", "hasAUMGreater100M", "typeEntity", "countryEntityOrganized", "countEmployeesTotal", "countEmployeesInvestmentAdvisory", "amountAUMTotal", "amountAUMDiscretionary", "amountAUMNonDiscretionary", "countAccountsDiscretionary", "countAccountsNonDiscretionary", "countAccountsTotal", "isManagerSecuritiesPortfolio", "hasFeeAUM", "hasFeeHourlyCharge", "hasFeeSubscription", "hasFeeFixed", "hasFeeCommission", "hasFeePerformance", "hasFeeOther", "typeFeeOther", "isBrokerDealer", "isBrokerDealerRepresentative", "isCommodityPoolOperator", "isFuturesMerchant", "isRealEstateBrokerDealerAgent", "isInsuranceBrokerAgent", "isBank", "isTrustCompany", "isRegisteredMunicipalAdviser", "isRegisteredSecuritySwapDealer", "isRegistredSecuritySwapParticipant", "isAccountingFirm", "isLawFirm", "isOtherFinancialProductSalesperson", "typeOtherFinancialProductSalesperson", "countEmployeesBrokerDealer", "countEmployeesStateRegisteredInvestmentAdviser", "countEmployeesStateRegisteredInvestmentAdviserMultipleEntities", "countEmployeesLicensedInsuranceAgents", "countEmployeesSolicitAdvisoryClients", "hasFelonyPleaConviction", "hasFelonyCharge", "hasMisdemeanorPleaConviction"))
}
}
\seealso{
Other IAPD: 
\code{\link{adv_managers_filings}()},
\code{\link{adv_managers_metadata}()},
\code{\link{adv_managers_periods_summaries}()},
\code{\link{finra_entities}()},
\code{\link{finra_people}()}

Other ADV: 
\code{\link{adv_managers_filings}()},
\code{\link{adv_managers_periods_summaries}()}

Other entity search: 
\code{\link{adv_managers_filings}()},
\code{\link{adv_managers_metadata}()},
\code{\link{adv_managers_periods_summaries}()},
\code{\link{data_rf_leis}()},
\code{\link{finra_entities}()},
\code{\link{nareit_entities}()},
\code{\link{reit_funds}()},
\code{\link{rf_sec_13F_companies}()},
\code{\link{sec_bankruptcies}()},
\code{\link{sec_broker_dealers}()},
\code{\link{sec_ciks}()},
\code{\link{sec_closed_end_funds}()},
\code{\link{sec_cusips}()},
\code{\link{sec_filer}()},
\code{\link{sec_filing_entities}()},
\code{\link{sec_investment_companies}()},
\code{\link{sec_money_market_funds}()},
\code{\link{sec_municipal_advisors}()},
\code{\link{sec_xbrl_periods}()},
\code{\link{securities_offerings}()},
\code{\link{us_public_companies}()},
\code{\link{ycombinator_alumni}()}

Other fund data: 
\code{\link{adv_managers_filings}()},
\code{\link{adv_managers_metadata}()},
\code{\link{adv_managers_periods_summaries}()},
\code{\link{reit_funds}()},
\code{\link{sec_ciks}()},
\code{\link{sec_closed_end_funds}()},
\code{\link{sec_money_market_funds}()}
}
\concept{ADV}
\concept{IAPD}
\concept{entity search}
\concept{fund data}
