% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adv_functions.R
\name{adv_managers_metadata}
\alias{adv_managers_metadata}
\title{IAPD registered meta data}
\usage{
adv_managers_metadata(entity_names = NULL, crd_ids = NULL,
  score_threshold = 0.2, return_message = T)
}
\arguments{
\item{entity_names}{vector of entities to search}

\item{crd_ids}{numric vector of crds to search}

\item{score_threshold}{matching score threshold for the search name
if \code{NULL} there is no threshold}

\item{return_message}{return a message}
}
\value{
a data frame
}
\description{
This function returns meta data
on any Investment Adviser Public Disclosure [IAPD]
filing manager.  This function can be used to discover
managers to power the
\code{\link{adv_managers_filing()}} function.
}
\examples{
adv_managers_metadata(entity_names = c('Divco', 'EJF'), score_threshold = .2)
adv_managers_metadata(entity_names = "Blackstone", score_threshold = NULL)
adv_managers_metadata(crd_ids = 173787)
}
\seealso{
Other IAPD: \code{\link{adv_managers_current_period_summary}},
  \code{\link{adv_managers_filings}},
  \code{\link{adv_managers_periods_summaries}},
  \code{\link{finra_entities}}, \code{\link{finra_people}}

Other FINRA: \code{\link{finra_entities}},
  \code{\link{finra_people}}

Other fund data: \code{\link{adv_managers_current_period_summary}},
  \code{\link{adv_managers_filings}},
  \code{\link{adv_managers_periods_summaries}},
  \code{\link{reit_funds}}, \code{\link{sec_ciks}},
  \code{\link{sec_closed_end_funds}},
  \code{\link{sec_money_market_funds}}

Other entity search: \code{\link{adv_managers_current_period_summary}},
  \code{\link{adv_managers_filings}},
  \code{\link{adv_managers_periods_summaries}},
  \code{\link{data_rf_leis}}, \code{\link{finra_entities}},
  \code{\link{nareit_entities}}, \code{\link{reit_funds}},
  \code{\link{rf_sec_13F_companies}},
  \code{\link{sec_bankruptcies}},
  \code{\link{sec_broker_dealers}}, \code{\link{sec_ciks}},
  \code{\link{sec_closed_end_funds}},
  \code{\link{sec_cusips}}, \code{\link{sec_filer}},
  \code{\link{sec_filing_entities}},
  \code{\link{sec_investment_companies}},
  \code{\link{sec_money_market_funds}},
  \code{\link{sec_municipal_advisors}},
  \code{\link{securities_offerings}},
  \code{\link{us_public_companies}},
  \code{\link{ycombinator_alumni}}
}
\concept{FINRA}
\concept{IAPD}
\concept{entity search}
\concept{fund data}
