% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sec_2.R
\name{build_address}
\alias{build_address}
\title{Build Address from tibble}
\usage{
build_address(
  data,
  address_search_slugs = c("^address", "^streetAddress", "^city", "^state",
    "^codeState", "^codeCountry", "^country", "^zipcode"),
  include_snake_versions = T,
  part_threshold = 3,
  snake_names = F,
  return_message = T
)
}
\arguments{
\item{data}{\code{tibble}}

\item{address_search_slugs}{vector of slugs identifying address features - defaults to `c("^address", "^streetAddress", "^city", "^state", "^codeState", "^codeCountry", "^country", "^zipcode")`}

\item{include_snake_versions}{`TRUE` includes snaked version of names}

\item{part_threshold}{minimum number of matches}

\item{snake_names}{if \code{TRUE} snakes names}

\item{return_message}{if \code{TRUE} returns a message}
}
\value{

}
\description{
Build Address from tibble
}
