% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterfall_functions.R
\name{calculate_cash_flow_waterfall}
\alias{calculate_cash_flow_waterfall}
\title{Get cash flow waterfall data for given periods and cash flows}
\usage{
calculate_cash_flow_waterfall(dates = c("2015-03-11", "2015-11-20",
  "2016-10-15"), cash_flows = c(-1e+05, -2e+05, 698906.76849),
  working_capital = 0, promote_structure = c("20 / 12", "30 / 18"),
  distribution_frequency = NA, is_actual_360 = T,
  widen_promote_structure = F, bind_to_cf = F, remove_zero_cols = T,
  widen_waterfall = F)
}
\arguments{
\item{dates}{A vector of dates in year-month-day format}

\item{cash_flows}{A vector of cash flows}

\item{working_capital}{Amount of working capital}

\item{promote_structure}{Vector of promte structures}

\item{is_actual_360}{Is the rate of return actual 360}
}
\description{
Get cash flow waterfall data for given periods and cash flows
}

