% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterfall_functions.R
\name{calculate_data_periods_irr}
\alias{calculate_data_periods_irr}
\title{Get an IRR for a set of cash flows}
\usage{
calculate_data_periods_irr(dates = c("2016-06-01", "2017-05-31", "2018-05-31",
  "2019-05-31", "2020-05-31", "2021-05-31", "2022-05-31", "2023-05-31",
  "2024-05-31", "2025-05-31", "2026-05-31"), cash_flows = c(-3000,
  478.515738547242, 478.515738547242, 478.515738547242, 478.515738547242,
  478.515738547242, 478.515738547242, 478.515738547242, 478.515738547242,
  478.515738547242, 478.515738547278), date_format = "\%Y-\%m-\%d",
  scale_to_100 = F, return_percentage = F, return_df = T,
  return_wide = T, return_message = T)
}
\arguments{
\item{dates}{A vector of dates}

\item{cash_flows}{A vector of cash flows}

\item{date_format}{Date formate}

\item{scale_to_100}{Scale to 100}

\item{return_percentage}{Return percentages}

\item{return_df}{Return a data frame}
}
\description{
Get an IRR for a set of cash flows
}

