% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterfall_functions.R
\name{calculate_days_accrued_pref}
\alias{calculate_days_accrued_pref}
\title{Calculate accrued preference for a stated period of days}
\usage{
calculate_days_accrued_pref(pct_pref = 0.1, is_actual_360 = T, days = 31,
  equity_bb = 1700000, pref_accrued_bb = 0)
}
\arguments{
\item{pref_accrued_bb}{}
}
\description{
Calculate accrued preference for a stated period of days
}

