% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bond_functions.R
\name{calculate_irr_bond_future}
\alias{calculate_irr_bond_future}
\title{Get bond data data frame}
\usage{
calculate_irr_bond_future(face_value = 1e+06, date_purchase = "2013-01-10",
  id_period_type = "S", assign_to_environment = T,
  date_issue = "2012-11-15", bond_duration_years = 10,
  purchase_price = "97-18+", bond_coupon = 1 + (5/8), use_currency = T,
  unnest_data = F)
}
\arguments{
\item{unnest_data}{}
}
\description{
Get bond data data frame
}
\examples{
calculate_irr_bond_future(face_value = 1000000, date_purchase = "2013-01-10", id_period_type = "S", date_issue = "2012-11-15", bond_duration_years = 10, purchase_price = "97-18+", bond_coupon =  1 + (5 / 8), use_currency = T)
}

