% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterfall_functions.R
\name{calculate_irr_periods}
\alias{calculate_irr_periods}
\title{Interal rate of return}
\usage{
calculate_irr_periods(dates = c("2016-06-01", "2017-05-31", "2018-05-31",
  "2019-05-31", "2020-05-31", "2021-05-31", "2022-05-31", "2023-05-31",
  "2024-05-31", "2025-05-31", "2026-05-31"), cash_flows = c(-3000,
  478.515738547242, 478.515738547242, 478.515738547242, 478.515738547242,
  478.515738547242, 478.515738547242, 478.515738547242, 478.515738547242,
  478.515738547242, 478.515738547278), date_format = "\%Y-\%m-\%d",
  scale_to_100 = FALSE, return_percentage = FALSE, return_df = TRUE,
  return_wide = TRUE, return_message = TRUE)
}
\arguments{
\item{dates}{vector of dates, year-month-date format}

\item{cash_flows}{vector of cash flows}

\item{date_format}{date format}

\item{scale_to_100}{\code{TRUE} scale to 100}

\item{return_percentage}{\code{TRUE} return percentages}

\item{return_df}{\code{TRUE} returns a data frame}

\item{return_message}{\code{TRUE} return a message after data import}
}
\value{
\code{data_frame}
}
\description{
This function returns a data frame that produces
the \href{https://en.wikipedia.org/wiki/Internal_rate_of_return}{internal rate of return}
for specified dates and cash flows
}
\examples{
calculate_irr_periods(dates = c("2016-06-01","2017-05-31", "2018-05-31", "2019-05-31", "2020-05-31", "2021-05-31",
"2022-05-31", "2023-05-31", "2024-05-31", "2025-05-31", "2026-05-31"), cash_flows = c( -3000, 478.515738547242, 478.515738547242, 478.515738547242, 478.515738547242,
478.515738547242, 478.515738547242, 478.515738547242, 478.515738547242, 478.515738547242, 478.515738547278 ), date_format = '\%Y-\%m-\%d', scale_to_100 = FALSE,
return_percentage = FALSE, return_df = TRUE, return_wide = TRUE, return_message = TRUE)
}
\seealso{
Other calculation: \code{\link{calculate_cash_flow_dates}},
  \code{\link{calculate_cash_flow_waterfall_partnership}},
  \code{\link{calculate_cash_flow_waterfall}},
  \code{\link{calculate_cash_flows_returns}},
  \code{\link{calculate_days_accrued_pref}},
  \code{\link{calculate_leverage_metrics}},
  \code{\link{calculate_loan_payment}},
  \code{\link{calculate_residual_valuation_cap_rates}},
  \code{\link{calculate_residual_valuation_ebitda_multiples}},
  \code{\link{calculate_share_proceeds}},
  \code{\link{calculate_valuation_post_money}},
  \code{\link{tidy_promote_structure}}

Other leveraged finance calculation: \code{\link{calculate_cash_flow_dates}},
  \code{\link{calculate_cash_flow_waterfall_partnership}},
  \code{\link{calculate_cash_flow_waterfall}},
  \code{\link{calculate_cash_flows_returns}},
  \code{\link{calculate_leverage_metrics}},
  \code{\link{calculate_loan_payment}}

Other partnership calculation: \code{\link{calculate_cash_flow_dates}},
  \code{\link{calculate_cash_flow_waterfall_partnership}},
  \code{\link{calculate_cash_flow_waterfall}},
  \code{\link{calculate_cash_flows_returns}},
  \code{\link{calculate_days_accrued_pref}},
  \code{\link{tidy_promote_structure}}
}
\concept{calculation}
\concept{leveraged finance calculation}
\concept{partnership calculation}
