% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debt_functions.R
\name{calculate_loan_payment}
\alias{calculate_loan_payment}
\title{Get's Loan Payment data}
\usage{
calculate_loan_payment(loan_start_date = "2016-06-01",
  amount_initial_draw = 3000, is_interest_only = F,
  interest_only_periods = 24, interest_rate = 10, is_actual_360 = T,
  amortization_years = 10, amortization_months = 0, term_years = 10,
  term_months = 0, pct_loan_fee = 0, balloon_year = 10,
  override_monthly_interest = F, interest_reserve_period = 0,
  balloon_month = 0, return_annual_summary = F)
}
\arguments{
\item{return_annual_summary}{}
}
\description{
Get's Loan Payment data
}

