% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf_functions.R
\name{calculate_residual_valuation_cap_rates}
\alias{calculate_residual_valuation_cap_rates}
\title{Calculate residual value}
\usage{
calculate_residual_valuation_cap_rates(cap_rates = c(0.05, 0.0525, 0.06, 0.2),
  net_operating_income = "$27,500,000", cost_of_sale = "5\%",
  debt_balance = "$350,000,000", return_wide = T)
}
\arguments{
\item{cap_rates}{Vector of Capitalization Rates in percent or character percent form}

\item{net_operating_income}{Vector of Net Operating Income in numeric or character numeric/currency form}

\item{cost_of_sale}{Vector of Cost of Sale in percent or character percent form}

\item{debt_balance}{Vector of anticipated Debt Balance at sale in numeric or character numeric/currency form}
}
\description{
Calculate residual value
}

