% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf_functions.R
\name{calculate_residual_valuation_ebitda_multiples}
\alias{calculate_residual_valuation_ebitda_multiples}
\title{Calculate residual value for a set of given EBITDA based inputs}
\usage{
calculate_residual_valuation_ebitda_multiples(ebitda_multiples = c(5, 10, 15,
  20), ebitda = "$27,500,000", cost_of_sale = "5\%",
  debt_balance = "$350,000,000", return_wide = T)
}
\arguments{
\item{ebitda_multiples}{Vector of EBITDA Multiples in numeric or character}

\item{ebitda}{Vector of EBITDA in numeric or character numeric/currency form}

\item{cost_of_sale}{Vector of Cost of Sale in percent or character percent form}

\item{debt_balance}{Vector of anticipated Debt Balance at sale in numeric or character numeric/currency form}

\item{return_wide}{Return data in wide or long form}
}
\description{
Calculate residual value for a set of given EBITDA based inputs
}

