% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf_functions.R
\name{calculate_ust_futures_conversion_factor}
\alias{calculate_ust_futures_conversion_factor}
\title{Get conversion factor for US treasury futures curve}
\usage{
calculate_ust_futures_conversion_factor(date_delivery = "2009-03-01",
  date_maturity = "2012-01-15", bond_coupon = 1.125, return_df = T)
}
\arguments{
\item{return_df}{}
}
\description{
Get conversion factor for US treasury futures curve
}

