% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf_functions.R
\name{calculate_valuation_post_money}
\alias{calculate_valuation_post_money}
\title{Calculates range of post investment valuations}
\usage{
calculate_valuation_post_money(pre_money_valuation = "$45,000,000",
  percent_sold = "10\%", return_wide = T)
}
\arguments{
\item{pre_money_valuation}{Vector of of valuations in numeric or character}

\item{percent_sold}{Vector of of amount of business sold in percent or character percent form}

\item{return_wide}{Return data in wide or long form}
}
\description{
Calculates range of post investment valuations
}

