% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venture_data_functions.R
\name{cb_unicorns}
\alias{cb_unicorns}
\title{CB Insight data}
\usage{
cb_unicorns(return_message = TRUE)
}
\arguments{
\item{return_message}{\code{TRUE} return a message after data import}
}
\value{
\code{tibble}
}
\description{
This function returns information about startups whose valuation
exceeds $1B as identified by CB Insights.
}
\examples{
cb_unicorns(return_message = TRUE)
}
\references{
\href{https://cbinsights/}{CB Insights}
}
\seealso{
Other venture capital: 
\code{\link{calculate_valuation_post_money}()},
\code{\link{ycombinator_alumni}()}
}
\concept{venture capital}
