% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sec_2.R
\name{cik_filing_counts}
\alias{cik_filing_counts}
\title{CIK Filing Counts}
\usage{
cik_filing_counts(cik, return_message = TRUE)
}
\arguments{
\item{cik}{CIK codes}

\item{return_message}{if \code{TRUE} returns a message}
}
\value{
A tibble containing filing counts for each CIK
}
\description{
CIK Filing Counts
}
\examples{
\dontrun{
cik_filing_counts(cik = 886982)
}
}
