% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_email}
\alias{clean_email}
\title{Clean a vector of email addresses}
\usage{
clean_email(x)
}
\arguments{
\item{x}{character vector}
}
\value{
character vector
}
\description{
Lowercases, trims whitespace and control chars, validates RFC-5322-lite
shape. Anything that fails validation becomes \code{NA}.
}
