% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_phone}
\alias{clean_phone}
\title{Clean a vector of phone numbers}
\usage{
clean_phone(x, default_cc = "1")
}
\arguments{
\item{x}{character vector}

\item{default_cc}{default country code (digits, no plus). Default \code{"1"}.}
}
\value{
character vector
}
\description{
Strips formatting, strips leading "1" on 11-digit US numbers, validates
length 7..15, returns E.164-ish \code{+CC#######} strings.
Empty/garbage/short values become \code{NA}. Default country code is "1".
}
\examples{
clean_phone(c("(212) 583-5000", "212.583.5000", "1-212-583-5000",
             "+44 1534 754 502", "N/A", "x"))
# -> c("+12125835000", "+12125835000", "+12125835000", "+441534754502", NA, NA)
}
