% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_urls}
\alias{clean_urls}
\title{Clean a character vector of URLs}
\usage{
clean_urls(x)
}
\arguments{
\item{x}{character vector (or anything coercible)}
}
\value{
character vector of cleaned URLs (length may be 0..length(x))
}
\description{
Lowercases scheme + host, adds \code{http://} to bare domains, strips
trailing slashes on bare-root URLs, drops empty / garbage entries
(\code{NONE}, \code{n/a}, scheme-only, etc.), dedupes case-insensitively.
Returns an empty character vector if no values survive.
}
\examples{
clean_urls(c("HTTP://WWW.FOO.COM", "http://www.foo.com/", "n/a", "bar.org"))
# -> c("http://www.foo.com", "http://bar.org")
}
