% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markit.R
\name{companies_tickers}
\alias{companies_tickers}
\title{Companies ticker symbols}
\usage{
companies_tickers(companies = c("Vornado", "Snap"),
  include_trades = TRUE, return_message = TRUE)
}
\arguments{
\item{companies}{vector of comapnies}

\item{include_trades}{if \code{TRUE} includes trade data}
}
\value{
a \code{date_frame}
}
\description{
Company ticker data
}
\references{
\href{http://markit.com}{MARKIT}
}
