% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crisp.R
\name{crsp_files}
\alias{crsp_files}
\title{CRSP file dictionary}
\usage{
crsp_files(include_summary_files = TRUE)
}
\arguments{
\item{include_summary_files}{Logical. If \code{TRUE} (default), includes summary files.}
}
\value{
A tibble containing CRSP file metadata with URLs and file names.
}
\description{
This function returns information about available
CRSP index data
}
\examples{
\dontrun{
crsp_files()
}
}
\references{
\href{http://crsp.com}{The Center for Research in Security Prices}
}
\seealso{
Other CRSP: 
\code{\link{crsp_indicies_constituents}()},
\code{\link{crsp_indicies_returns}()}
}
\concept{CRSP}
