% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crisp.R
\name{crsp_indicies_constituents}
\alias{crsp_indicies_constituents}
\title{CRSP index constituents}
\usage{
crsp_indicies_constituents(return_message = TRUE)
}
\arguments{
\item{return_message}{Logical. If \code{TRUE}, returns progress messages.}
}
\value{
A tibble containing CRSP index constituent data including ticker,
  company name, and weight.
}
\description{
This function returns information
about a CRSP index's constituents
}
\examples{
\dontrun{
crsp_indicies_constituents(return_message = TRUE)
}
}
\references{
\href{http://crsp.com}{The Center for Research in Security Prices}
}
\seealso{
Other CRSP: 
\code{\link{crsp_files}()},
\code{\link{crsp_indicies_returns}()}

Other index constituents: 
\code{\link{msci_indicies_constituents}()},
\code{\link{nareit_constituent_years}()},
\code{\link{sp500_constituents}()}
}
\concept{CRSP}
\concept{index constituents}
