% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgar_rf.R
\name{data_location_codes}
\alias{data_location_codes}
\title{Locations}
\usage{
data_location_codes()
}
\value{
a \code{tibble}
}
\description{
This function returns location codes
and the name the location
}
\examples{
data_location_codes()
}
\seealso{
Other dictionary: \code{\link{data_sec_rules}},
  \code{\link{data_sic_naics_codes}},
  \code{\link{dictionary_fred_ids}},
  \code{\link{dictionary_sec_filing_codes}},
  \code{\link{dictionary_sec_form_codes}},
  \code{\link{dictionary_sic_codes}}
}
\concept{dictionary}
