% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgar_rf.R
\name{data_rf_us_tickers}
\alias{data_rf_us_tickers}
\title{Get United States publicly traded ticker data}
\usage{
data_rf_us_tickers(return_message = TRUE)
}
\arguments{
\item{return_message}{}
}
\description{
Get United States publicly traded ticker data
}
\examples{
data_us_tickers(return_message = TRUE)
}
