% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgar_rf.R
\name{data_sec_rules}
\alias{data_sec_rules}
\title{SEC Prefix Laws}
\usage{
data_sec_rules()
}
\value{
a  \code{data_frame}
}
\description{
This function returns prefix
descriptions of SEC rules and the
law that created the rule
}
\seealso{
Other dictionary: \code{\link{data_location_codes}},
  \code{\link{data_sic_naics_codes}},
  \code{\link{dictionary_fred_ids}},
  \code{\link{dictionary_sec_filing_codes}},
  \code{\link{dictionary_sec_form_codes}},
  \code{\link{dictionary_sic_codes}}
}
\concept{SEC
data_sec_rules()}
\concept{dictionary}
