% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgar_rf.R
\name{data_sic_naics_codes}
\alias{data_sic_naics_codes}
\title{SIC Codes}
\usage{
data_sic_naics_codes(filter_duplicates = TRUE)
}
\arguments{
\item{filter_duplicates}{\code{TRUE} removes duplicate entries}
}
\value{
a \code{tibble}
}
\description{
This function returns Standard Industrial Classification [SIC]
codes and the corresponding
North American Industry Classification System [NAICS] identification.
}
\examples{
data_sic_naics_codes(filter_duplicates = TRUE)
}
\seealso{
Other dictionary: 
\code{\link{data_location_codes}()},
\code{\link{data_sec_rules}()},
\code{\link{dictionary_fred_ids}()},
\code{\link{dictionary_sec_filing_codes}()},
\code{\link{dictionary_sec_form_codes}()},
\code{\link{dictionary_sic_codes}()}
}
\concept{dictionary}
