% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finbox.R
\name{dictionary_finbox_companies}
\alias{dictionary_finbox_companies}
\title{FinboxIO Companies}
\usage{
dictionary_finbox_companies(return_message = TRUE)
}
\arguments{
\item{return_message}{if \code{TRUE} return a message}
}
\value{
a \code{tibble} with company data including tickers, exchange info, and sectors
}
\description{
Returns companies with FinBOXIO data
}
\examples{
dictionary_finbox_companies()
}
