% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgar_rf.R
\name{dictionary_form_d_categories}
\alias{dictionary_form_d_categories}
\title{Form-D dictionary}
\usage{
dictionary_form_d_categories()

dictionary_form_d_categories()

dictionary_form_d_categories()
}
\value{
a \code{tibble}

a \code{tibble}

a \code{tibble}
}
\description{
This function returns searchable
industries for parsed SEC Form-D
filings

This function returns searchable
industries for parsed SEC Form-D
filings

This function returns searchable
industries for parsed SEC Form-D
filings
}
\examples{
dictionary_form_d_categories()
dictionary_form_d_categories()
dictionary_form_d_categories()
}
