% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fred_api.R
\name{dictionary_fred_ids}
\alias{dictionary_fred_ids}
\title{FRED Series IDs}
\usage{
dictionary_fred_ids(fred_file = NULL, return_message = TRUE)
}
\arguments{
\item{fred_file}{the type of dictionary you wish to import \itemize{
\item \code{NULL, All}:  returns all dictionary data in a nested data frame (default)
\item \code{series}: returns a data frame of the series name, id, and attributes
\item \code{tags}: returns a data frame of the series id and their corresponding search tag
\item \code{subindicies}: returns a data frame with the parent series and any sub-indicies if they have them
}}

\item{return_message}{\code{TRUE} return a message after data import}
}
\value{
a \code{tibble}
}
\description{
This function imports dictionaries for all possible FRED
series.  You can us this data to find series to search in the
\code{\link{fred_symbols}} function.
}
\details{
This dictionary was indexed on January 22, 2017, please be aware
that there may be new series added to FRED since then.
}
\examples{

dictionary_fred_ids(fred_file = NULL, return_message  = TRUE)
}
\seealso{
Other dictionary: 
\code{\link{dictionary_sec_filing_codes}()},
\code{\link{dictionary_sec_form_codes}()},
\code{\link{dictionary_sic_codes}()},
\code{\link{location_codes}()},
\code{\link{sec_rules}()},
\code{\link{sic_naics_codes}()}
}
\concept{FRED}
\concept{dictionary}
