% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgar_rf.R
\name{dictionary_sec_filing_codes}
\alias{dictionary_sec_filing_codes}
\title{SEC filing code dictionary}
\usage{
dictionary_sec_filing_codes()

dictionary_sec_filing_codes()

dictionary_sec_filing_codes()
}
\value{
a \code{tibble}

a \code{tibble}

a \code{tibble}
}
\description{
This function returns a
dictionary of SEC form filing types

This function returns a
dictionary of SEC form filing types

This function returns a
dictionary of SEC form filing types
}
\examples{
dictionary_sec_filing_codes()
dictionary_sec_filing_codes()
dictionary_sec_filing_codes()
}
\seealso{
Other SEC: 
\code{\link{data_rf_leis}()},
\code{\link{dictionary_sec_form_codes}()},
\code{\link{dictionary_sic_codes}()},
\code{\link{edgar_filing_streams}()},
\code{\link{mmf_owned_debt_securities}()},
\code{\link{recent_insider_trades}()},
\code{\link{rf_sec_13F_companies}()},
\code{\link{sec_bankruptcies}()},
\code{\link{sec_broker_dealers}()},
\code{\link{sec_ciks}()},
\code{\link{sec_closed_end_funds}()},
\code{\link{sec_cusips}()},
\code{\link{sec_failed_to_deliver_securities}()},
\code{\link{sec_filer}()},
\code{\link{sec_filing_entities}()},
\code{\link{sec_foia_requests}()},
\code{\link{sec_investment_companies}()},
\code{\link{sec_money_market_funds}()},
\code{\link{sec_municipal_advisors}()},
\code{\link{sec_securities_filing_counts}()},
\code{\link{sec_securities_metrics_by_exchange}()},
\code{\link{sec_xbrl_periods}()},
\code{\link{securities_offerings}()},
\code{\link{us_public_companies}()}

Other dictionary: 
\code{\link{data_location_codes}()},
\code{\link{data_sec_rules}()},
\code{\link{data_sic_naics_codes}()},
\code{\link{dictionary_fred_ids}()},
\code{\link{dictionary_sec_form_codes}()},
\code{\link{dictionary_sic_codes}()}

Other SEC: 
\code{\link{data_rf_leis}()},
\code{\link{dictionary_sec_form_codes}()},
\code{\link{dictionary_sic_codes}()},
\code{\link{edgar_filing_streams}()},
\code{\link{mmf_owned_debt_securities}()},
\code{\link{recent_insider_trades}()},
\code{\link{rf_sec_13F_companies}()},
\code{\link{sec_bankruptcies}()},
\code{\link{sec_broker_dealers}()},
\code{\link{sec_ciks}()},
\code{\link{sec_closed_end_funds}()},
\code{\link{sec_cusips}()},
\code{\link{sec_failed_to_deliver_securities}()},
\code{\link{sec_filer}()},
\code{\link{sec_filing_entities}()},
\code{\link{sec_foia_requests}()},
\code{\link{sec_investment_companies}()},
\code{\link{sec_money_market_funds}()},
\code{\link{sec_municipal_advisors}()},
\code{\link{sec_securities_filing_counts}()},
\code{\link{sec_securities_metrics_by_exchange}()},
\code{\link{sec_xbrl_periods}()},
\code{\link{securities_offerings}()},
\code{\link{us_public_companies}()}

Other dictionary: 
\code{\link{data_location_codes}()},
\code{\link{data_sec_rules}()},
\code{\link{data_sic_naics_codes}()},
\code{\link{dictionary_fred_ids}()},
\code{\link{dictionary_sec_form_codes}()},
\code{\link{dictionary_sic_codes}()}

Other SEC: 
\code{\link{data_rf_leis}()},
\code{\link{dictionary_sec_form_codes}()},
\code{\link{dictionary_sic_codes}()},
\code{\link{edgar_filing_streams}()},
\code{\link{mmf_owned_debt_securities}()},
\code{\link{recent_insider_trades}()},
\code{\link{rf_sec_13F_companies}()},
\code{\link{sec_bankruptcies}()},
\code{\link{sec_broker_dealers}()},
\code{\link{sec_ciks}()},
\code{\link{sec_closed_end_funds}()},
\code{\link{sec_cusips}()},
\code{\link{sec_failed_to_deliver_securities}()},
\code{\link{sec_filer}()},
\code{\link{sec_filing_entities}()},
\code{\link{sec_foia_requests}()},
\code{\link{sec_investment_companies}()},
\code{\link{sec_money_market_funds}()},
\code{\link{sec_municipal_advisors}()},
\code{\link{sec_securities_filing_counts}()},
\code{\link{sec_securities_metrics_by_exchange}()},
\code{\link{sec_xbrl_periods}()},
\code{\link{securities_offerings}()},
\code{\link{us_public_companies}()}

Other dictionary: 
\code{\link{data_location_codes}()},
\code{\link{data_sec_rules}()},
\code{\link{data_sic_naics_codes}()},
\code{\link{dictionary_fred_ids}()},
\code{\link{dictionary_sec_form_codes}()},
\code{\link{dictionary_sic_codes}()}
}
\concept{SEC}
\concept{dictionary}
