% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treasury_auctions.R
\name{dictionary_treasury_auction_urls}
\alias{dictionary_treasury_auction_urls}
\title{Treasury Auction URL Dictionary}
\usage{
dictionary_treasury_auction_urls()
}
\description{
Treasury Auction URL Dictionary
}
