% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ferrara_messaging.R
\name{.fm_danger}
\alias{.fm_danger}
\title{Styled danger/error message - The Fall}
\usage{
.fm_danger(...)
}
\arguments{
\item{...}{Message components}
}
\description{
Styled danger/error message - The Fall
}
\keyword{internal}
