% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ferrara_messaging.R
\name{.fm_divider}
\alias{.fm_divider}
\title{Section divider - The Scene Break}
\usage{
.fm_divider(title = NULL)
}
\arguments{
\item{title}{Optional section title}
}
\description{
Section divider - The Scene Break
}
\keyword{internal}
