% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ferrara_messaging.R
\name{.fm_entity_summary}
\alias{.fm_entity_summary}
\title{Entity summary message}
\usage{
.fm_entity_summary(entity_type, name, highlights = list())
}
\arguments{
\item{entity_type}{Type of entity (company, fund, etc.)}

\item{name}{Entity name}

\item{highlights}{Named list of key stats to display}
}
\description{
For functions that return entity data with highlights
}
\keyword{internal}
