% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ferrara_messaging.R
\name{.fm_financial_result}
\alias{.fm_financial_result}
\title{Financial summary message}
\usage{
.fm_financial_result(
  metric_name,
  value,
  is_currency = FALSE,
  is_percent = FALSE
)
}
\arguments{
\item{metric_name}{Name of the financial metric}

\item{value}{The value (will be formatted)}

\item{is_currency}{Whether to format as currency}

\item{is_percent}{Whether to format as percentage}
}
\description{
Financial summary message
}
\keyword{internal}
