% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ferrara_messaging.R
\name{.fm_headline}
\alias{.fm_headline}
\title{Styled headline - The Title Card}
\usage{
.fm_headline(...)
}
\arguments{
\item{...}{Message components}
}
\description{
Styled headline - The Title Card
}
\keyword{internal}
