% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ferrara_messaging.R
\name{.fm_info}
\alias{.fm_info}
\title{Styled info message - The Surveillance}
\usage{
.fm_info(...)
}
\arguments{
\item{...}{Message components}
}
\description{
Styled info message - The Surveillance
}
\keyword{internal}
