% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ferrara_messaging.R
\name{.fm_spotlight}
\alias{.fm_spotlight}
\title{Random spotlight message (for YC, unicorns, etc.)}
\usage{
.fm_spotlight(company, description = NULL, url = NULL, extra = list())
}
\arguments{
\item{company}{Company name}

\item{description}{Company description}

\item{url}{Optional URL}

\item{extra}{Named list of additional fields}
}
\description{
Random spotlight message (for YC, unicorns, etc.)
}
\keyword{internal}
