% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ferrara_messaging.R
\name{.fm_success}
\alias{.fm_success}
\title{Styled success message - The Deal Closes}
\usage{
.fm_success(...)
}
\arguments{
\item{...}{Message components (glue-style interpolation supported)}
}
\description{
Styled success message - The Deal Closes
}
\examples{
\dontrun{
.fm_success("Acquired {.fm_value 1500} records from {.fm_entity SEC EDGAR}")
}
}
\keyword{internal}
