% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ferrara_messaging.R
\name{.fm_time_range}
\alias{.fm_time_range}
\title{Time range message}
\usage{
.fm_time_range(start_year, end_year, n_records, data_type = "records")
}
\arguments{
\item{start_year}{Start year}

\item{end_year}{End year}

\item{n_records}{Number of records}

\item{data_type}{Type of data}
}
\description{
Time range message
}
\keyword{internal}
