% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ferrara_messaging.R
\name{.fm_warning}
\alias{.fm_warning}
\title{Styled warning message - The Danger}
\usage{
.fm_warning(...)
}
\arguments{
\item{...}{Message components}
}
\description{
Styled warning message - The Danger
}
\keyword{internal}
