% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unified_securities.R
\name{.scan_tv_paged}
\alias{.scan_tv_paged}
\title{Paginate a scanner call in 50k chunks up to max_rows}
\usage{
.scan_tv_paged(
  endpoint,
  cols,
  filter = NULL,
  max_rows = 50000L,
  page_size = 50000L,
  return_message = TRUE
)
}
\description{
Paginate a scanner call in 50k chunks up to max_rows
}
\keyword{internal}
