% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{drop_na_columns}
\alias{drop_na_columns}
\title{Drops NA columns}
\usage{
drop_na_columns(data)
}
\arguments{
\item{data}{}
}
\description{
Drops NA columns
}
\examples{
data_frame(nameFirm = 'Goldman Sachs', countSuperHeros = NA, countCriminals = 0, countFinedEmployees = 10) \%>\% drop_na_columns()
}

