% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sec_2.R
\name{edgar_entities_cik}
\alias{edgar_entities_cik}
\title{SEC registered entity search}
\usage{
edgar_entities_cik(search_names, nest_data = FALSE, return_message = TRUE)
}
\arguments{
\item{search_names}{vector of entity names to search}

\item{nest_data}{if \code{TRUE} nests data}

\item{return_message}{if \code{TRUE} returns a message}
}
\value{
A tibble containing CIK codes for matching entities
}
\description{
SEC registered entity search
}
\examples{
\dontrun{
edgar_entities_cik(c("Rockwood", "BREA", "EJF"))
}
}
