% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgar_rf.R
\name{edgar_filing_streams}
\alias{edgar_filing_streams}
\title{SEC filing streams}
\usage{
edgar_filing_streams(start_date = "2017-02-15", end_date = Sys.Date(),
  only_most_recent_data = FALSE, index_type = "master",
  table_name_initial = "Filing Logs", parse_all_filings = FALSE,
  parse_complete_text_filings = FALSE, parse_form_d = FALSE,
  parse_13F = FALSE, parse_small_offerings = FALSE,
  parse_form_3_4s = FALSE, parse_asset_files = FALSE,
  parse_xbrl = FALSE, assign_to_environment = TRUE, nest_data = TRUE,
  return_message = TRUE)
}
\arguments{
\item{start_date}{starting date in year-month-date form}

\item{end_date}{ending date starting in year-month-date form}

\item{only_most_recent_data}{\code{TRUE} return only most recent day's filing stream}

\item{index_type}{type of index to parse \itemize{
\item \code{master}: parses master log (default)
\item \code{compamy}: parses company log
\item \code{filer}: parses filer log
}}

\item{nest_data}{\code{TRUE} return nested data frame}

\item{return_message}{\code{TRUE} return a message after data import}
}
\value{
nested \code{tibble} or \code{tibble} if \code{nest_data = FALSE}
}
\description{
This function parses daily
SEC filing log starting in 1994
for specified periods
}
\examples{
\dontrun{
edgar_filing_streams(start_date = "2016-01-01",
end_date = Sys.Date(), only_most_recent_data = FALSE, index_type = 'master',
nest_data = TRUE,
return_message = TRUE)
}
}
\references{
\href{http://sec.gov}{The Securities and Exchange Commission}
}
\seealso{
Other SEC: \code{\link{data_rf_leis}},
  \code{\link{dictionary_sec_filing_codes}},
  \code{\link{dictionary_sec_form_codes}},
  \code{\link{dictionary_sic_codes}},
  \code{\link{mmf_owned_debt_securities}},
  \code{\link{recent_insider_trades}},
  \code{\link{rf_sec_13F_companies}},
  \code{\link{sec_bankruptcies}},
  \code{\link{sec_broker_dealers}}, \code{\link{sec_ciks}},
  \code{\link{sec_closed_end_funds}},
  \code{\link{sec_cusips}},
  \code{\link{sec_failed_to_deliver_securities}},
  \code{\link{sec_filer}},
  \code{\link{sec_filing_entities}},
  \code{\link{sec_foia_requests}},
  \code{\link{sec_investment_companies}},
  \code{\link{sec_money_market_funds}},
  \code{\link{sec_municipal_advisors}},
  \code{\link{sec_securities_filing_counts}},
  \code{\link{sec_securities_metrics_by_exchange}},
  \code{\link{sec_xbrl_periods}},
  \code{\link{securities_offerings}},
  \code{\link{us_public_companies}}
}
\concept{SEC}
\concept{filing search}
