% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sec_2.R
\name{edgar_recent_filings}
\alias{edgar_recent_filings}
\title{Most recent EDGAR filings by type}
\usage{
edgar_recent_filings(
  forms = c("All", "10-D", "10-K"),
  table_name_initial = "Recent Filings",
  parse_all_filings = TRUE,
  parse_form_d = FALSE,
  parse_complete_text_filings = FALSE,
  parse_13F = FALSE,
  parse_small_offerings = FALSE,
  parse_form_3_4s = FALSE,
  parse_asset_files = FALSE,
  parse_xbrl = FALSE,
  assign_to_environment = TRUE,
  nest_data = FALSE,
  return_message = TRUE
)
}
\arguments{
\item{forms}{vector of form types to retrieve}

\item{table_name_initial}{initial table name for assignment}

\item{parse_all_filings}{if \code{TRUE} parses all filings}

\item{parse_form_d}{if \code{TRUE} parses Form D filings}

\item{parse_complete_text_filings}{if \code{TRUE} parses complete text filings}

\item{nest_data}{if \code{TRUE} nests data}

\item{return_message}{if \code{TRUE} returns a message}
}
\value{
A tibble containing recent EDGAR filings
}
\description{
Most recent EDGAR filings by type
}
\examples{
\dontrun{
edgar_recent_filings(forms = c("10-K", "10-Q"))
}
}
