% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sec_2.R
\name{edgar_search_terms}
\alias{edgar_search_terms}
\title{EDGAR Search for Terms}
\usage{
edgar_search_terms(
  search_terms = NULL,
  parameter = NULL,
  year_start = NULL,
  year_end = NULL,
  table_name_initial = "All Filings",
  parse_all_filings = TRUE,
  parse_complete_text_filings = FALSE,
  parse_form_d = FALSE,
  parse_13F = FALSE,
  parse_small_offerings = FALSE,
  parse_form_3_4s = FALSE,
  parse_asset_files = FALSE,
  parse_xbrl = FALSE,
  assign_to_environment = TRUE,
  nest_data = TRUE,
  return_message = TRUE
)
}
\arguments{
\item{search_terms}{vector of search terms}

\item{parameter}{additional search parameters}

\item{year_start}{starting year for search}

\item{year_end}{ending year for search}

\item{table_name_initial}{initial table name for assignment}

\item{parse_all_filings}{Logical. If \code{TRUE}, parses all filings.}

\item{parse_complete_text_filings}{Logical. If \code{TRUE}, parses complete text of filings.}

\item{parse_form_d}{Logical. If \code{TRUE}, parses Form D filings.}

\item{parse_13F}{Logical. If \code{TRUE}, parses 13F filings.}

\item{parse_small_offerings}{Logical. If \code{TRUE}, parses small offerings.}

\item{parse_form_3_4s}{Logical. If \code{TRUE}, parses Form 3 and 4 filings.}

\item{parse_asset_files}{Logical. If \code{TRUE}, parses asset files.}

\item{parse_xbrl}{Logical. If \code{TRUE}, parses XBRL data.}

\item{assign_to_environment}{Logical. If \code{TRUE}, assigns results to environment.}

\item{nest_data}{Logical. If \code{TRUE}, nests data.}

\item{return_message}{Logical. If \code{TRUE}, returns progress messages.}
}
\value{
A tibble containing EDGAR filings matching the search terms.
}
\description{
EDGAR Search for Terms
}
\examples{
\dontrun{
edgar_search_terms(search_terms = "China", year_start = 2020)
}
}
