% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sec_2.R
\name{edgar_tickers}
\alias{edgar_tickers}
\title{SEC listed public companys}
\usage{
edgar_tickers(
  include_ticker_information = FALSE,
  join_sic = TRUE,
  snake_names = TRUE,
  return_message = TRUE
)
}
\arguments{
\item{include_ticker_information}{if \code{TRUE} returns ticker information}

\item{join_sic}{if \code{TRUE} joins SIC code data}

\item{snake_names}{if \code{TRUE} returns snake case names}

\item{return_message}{if \code{TRUE} returns a message}
}
\value{
A tibble containing SEC listed public company ticker data
}
\description{
SEC listed public companys
}
\examples{
edgar_tickers()
}
\seealso{
Other SEC EDGAR: 
\code{\link{sec_13f_cusips}()},
\code{\link{sec_13f_holdings_bulk}()},
\code{\link{sec_cik_master}()},
\code{\link{sec_edgar_quarterly_index}()},
\code{\link{sec_fails_to_deliver}()},
\code{\link{sec_financial_statement_data_sets}()},
\code{\link{sec_nport_bulk}()},
\code{\link{sec_tickers_exchange}()}
}
\concept{SEC EDGAR}
