% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adv_functions.R
\name{find_text_node}
\alias{find_text_node}
\title{Title}
\usage{
find_text_node(node_text, hit_words = "Total Number of Clients",
  off_set = 4, is_numeric_node = T)
}
\arguments{
\item{is_numeric_node}{}
}
\description{
Title
}

