% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adv_functions.R
\name{finra_entities}
\alias{finra_entities}
\title{FINRA registered entities}
\usage{
finra_entities(
  entity_names = NULL,
  ocr_pdf = TRUE,
  score_threshold = 0.2,
  return_message = TRUE
)
}
\arguments{
\item{entity_names}{vector of names to search}

\item{ocr_pdf}{if \code{TRUE} returns OCR'd broker report PDF}

\item{score_threshold}{matching score threshold for the search name
if \code{NULL} there is no threshold}

\item{return_message}{return a message upon parsing \code{TRUE, FALS}}
}
\value{
a \code{data frame}
}
\description{
This function returns information for any
Financial Industry Regulatory Authority [FINRA] registered
entity.
}
\note{
Use \code{\link{finra_people}} for registered people
}
\examples{
finra_entities(entity_names = c("EJF", "Blackstone", "Next Play", "Simple Capital"), score_threshold = .25,
ocr_pdf = TRUE)
}
\references{
\href{http://www.finra.org/}{FINRA}
}
\seealso{
Other FINRA: 
\code{\link{adv_managers_metadata}()},
\code{\link{finra_people}()}

Other IAPD: 
\code{\link{adv_managers_current_period_summary}()},
\code{\link{adv_managers_filings}()},
\code{\link{adv_managers_metadata}()},
\code{\link{adv_managers_periods_summaries}()},
\code{\link{finra_people}()}

Other entity search: 
\code{\link{adv_managers_current_period_summary}()},
\code{\link{adv_managers_filings}()},
\code{\link{adv_managers_metadata}()},
\code{\link{adv_managers_periods_summaries}()},
\code{\link{nareit_entities}()},
\code{\link{reit_funds}()},
\code{\link{rf_leis}()},
\code{\link{rf_sec_13F_companies}()},
\code{\link{sec_bankruptcies}()},
\code{\link{sec_broker_dealers}()},
\code{\link{sec_ciks}()},
\code{\link{sec_closed_end_funds}()},
\code{\link{sec_cusips}()},
\code{\link{sec_filer}()},
\code{\link{sec_filing_entities}()},
\code{\link{sec_investment_companies}()},
\code{\link{sec_money_market_funds}()},
\code{\link{sec_municipal_advisors}()},
\code{\link{sec_xbrl_periods}()},
\code{\link{securities_offerings}()},
\code{\link{us_public_companies}()},
\code{\link{ycombinator_alumni}()}
}
\concept{FINRA}
\concept{IAPD}
\concept{entity search}
