% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finra.R
\name{finra_people}
\alias{finra_people}
\title{FINRA registered people}
\usage{
finra_people(search_name = NULL, ocr_pdf = TRUE, return_message = TRUE)
}
\arguments{
\item{search_name}{vector of names to search}

\item{ocr_pdf}{Logical. If \code{TRUE}, returns OCR'd broker report PDF.}

\item{return_message}{Logical. If \code{TRUE}, returns progress messages.}
}
\value{
A tibble with FINRA registered person information.
}
\description{
This function returns information for any
Financial Industry Regulatory Authority [FINRA] registered
person.
}
\note{
Use \code{\link{finra_entities}} for registered entities
}
\examples{
\dontrun{
finra_people(search_name = 'Lloyd Blankfein', ocr_pdf = TRUE)
}
}
\references{
\href{http://www.finra.org/}{FINRA}
}
\seealso{
Other FINRA: 
\code{\link{adv_managers_metadata}()},
\code{\link{finra_entities}()}

Other IAPD: 
\code{\link{adv_managers_current_period_summary}()},
\code{\link{adv_managers_filings}()},
\code{\link{adv_managers_metadata}()},
\code{\link{adv_managers_periods_summaries}()},
\code{\link{finra_entities}()}
}
\concept{FINRA}
\concept{IAPD}
\concept{person search}
