% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fred_api.R
\name{fred_symbols}
\alias{fred_symbols}
\title{Federal Reserve Economic Data time series tibble}
\usage{
fred_symbols(symbols = c("DGS2", "DGS10", "DGS30"),
  transformations = c("default"), widen_data = F,
  convert_date_time = TRUE, nest_data = TRUE,
  include_metadata = TRUE, return_message = TRUE)
}
\arguments{
\item{symbols}{fred symbols to search, see \href{https://fred.stlouisfed.org/tags/series}{FRED symbols} for options}

\item{transformations}{transformations}

\item{widen_data}{if \code{TRUE} widens data}

\item{convert_date_time}{converts date from datetime to date form}

\item{nest_data}{\code{TRUE} return nested data frame}

\item{include_metadata}{if \code{TRUE} returns meta data from time series}

\item{return_message}{if \code{TRUE} return message}
}
\value{
a \code{tibble}
}
\description{
This function returns a data for specified series from
 \href{https://fred.stlouisfed.org/}{FRED}.
}
\examples{
fred_symbols(symbols = c('DGS30','DGS10','DGS2'),
return_wide = TRUE, nest_data = FALSE)

fred_symbols(
symbols = c("CPIAUCSL", "A191RL1Q225SBEA", "UNRATE"),
convert_date_time = TRUE,
nest_data = TRUE,
include_metadata = TRUE
)

}
\seealso{
Other index values: \code{\link{crsp_indicies_returns}},
  \code{\link{msci_realtime_index_values}},
  \code{\link{nareit_annual_subsector_returns}},
  \code{\link{nareit_monthly_returns}}
}
\concept{index values}
