% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fred_api.R
\name{fred_terms_ids}
\alias{fred_terms_ids}
\title{FRED Terms Series IDs}
\usage{
fred_terms_ids(
  search_terms = NULL,
  use_json_api = TRUE,
  snake_names = FALSE,
  return_message = TRUE
)
}
\arguments{
\item{search_terms}{vector of search terms}

\item{use_json_api}{Logical. If \code{TRUE} (default), uses faster JSON API
but results limited to 100.}

\item{snake_names}{Logical. If \code{TRUE}, converts column names to snake_case.}

\item{return_message}{Logical. If \code{TRUE}, returns progress messages.}
}
\value{
A tibble containing matching FRED series IDs and metadata.
}
\description{
This function returns
matching FRED series ID for a given
search term.  When using
`use_json_api` search results limited to 100.
}
\examples{
\dontrun{
fred_terms_ids(search_terms = c("China Debt", "China Housing"))
}
}
