% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trading_view.R
\name{generate_tv_bond_query}
\alias{generate_tv_bond_query}
\title{Default Bond query metric}
\usage{
generate_tv_bond_query(
  column_names = c("country_code", "name", "coupon", "maturity_date", "close", "change",
    "change_abs", "high", "low", "Recommend.All", "description", "subtype", "pricescale",
    "minmov", "fractional", "minmove2", "ask", "all_time_low", "sector", "country2",
    "bid", "country", "current_session", "all_time_high", "rtc", "type", "expiration",
    "exchange", "pre_change", "pre_change_abs", "change_from_open_abs",
    "change_from_open", "post_change", "time", "open", "gap")
)
}
\arguments{
\item{column_names}{if not \code{NULL} vector of column names}
}
\description{
Default Bond query metric
}
