% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_class_df}
\alias{get_class_df}
\title{Class DF}
\usage{
get_class_df(data)
}
\arguments{
\item{data}{a \code{tibble}}
}
\value{
\code{tibble}
}
\description{
This function returns the
column classes of a specified
data frame.
}
\examples{
get_class_df(mtcars)
}
\seealso{
Other utility function: \code{\link{drop_na_columns}},
  \code{\link{tidy_column_formats}},
  \code{\link{tidy_column_relations}}
}
\concept{utility function}
