% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adv_functions.R
\name{get_data_adv_managers_current_period_summary}
\alias{get_data_adv_managers_current_period_summary}
\title{Get ADV filing data for the most recent filing period}
\usage{
get_data_adv_managers_current_period_summary(file_directory = NULL,
  folder_name = "adv_data", remove_files = F, empty_trash = F)
}
\arguments{
\item{file_directory}{Location of the directory you want to save your data into, if none specified a temporary file will be created}

\item{folder_name}{Name of the folder you want the data to be downloaded into}

\item{remove_files}{Remove the files from the folders}

\item{empty_trash}{Do you wish to empty the trash after being read into R}
}
\description{
Get ADV filing data for the most recent filing period
}
\examples{
get_data_adv_managers_current_period_summary(ile_directory = NULL, folder_name = 'adv_data', remove_files = T, empty_trash = T)
}

