% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adv_functions.R
\name{get_data_adv_managers_filings}
\alias{get_data_adv_managers_filings}
\title{Get detailed ADV form data for specified search name or CRD id}
\usage{
get_data_adv_managers_filings(search_names = NULL, crd_ids = NULL,
  all_sections = T, section_names = c("Registration",
  "Identifying Information", "Organization", "Successions",
  "Private Fund Reporting", "Direct Manager Owners", "Indirect Manager Owners",
  "Other Manager Information", "Manager Signatories"), flatten_tables = T,
  gather_data = F, assign_to_environment = T)
}
\arguments{
\item{search_names}{Names of the entities you want to search}

\item{crd_ids}{CRD ids you want to search}

\item{all_sections}{Do you want to search all ADV Sections}

\item{section_names}{If not all sections, which sections}

\item{flatten_tables}{Do you want the data the data with singular values flattened into a single data frame}

\item{gather_data}{Do you want the data in gathered form}

\item{assign_to_environment}{Do you want to save the invidual data frames to your global environment}
}
\description{
Get detailed ADV form data for specified search name or CRD id
}
\examples{
get_data_adv_managers_filings(search_names = c('Blackstone Real Estate'), crd_ids = NULL, all_sections = T,  section_names = NULL, flatten_tables = T, gather_data = F, assign_to_environment = T)
}

