% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adv_functions.R
\name{get_data_adv_managers_summary_periods}
\alias{get_data_adv_managers_summary_periods}
\title{Get ADV summary data for specified periods and filing types}
\usage{
get_data_adv_managers_summary_periods(periods = c("2016-08"),
  all_periods = T, is_exempt = F, only_most_recent = F,
  file_directory = "Desktop/adv_data", remove_files = T, empty_trash = T)
}
\description{
Get ADV summary data for specified periods and filing types
}

