% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adv_functions.R
\name{get_data_adv_period_urls}
\alias{get_data_adv_period_urls}
\title{Get ADV Summary Filing Data URLs for All Applicable Periods}
\usage{
get_data_adv_period_urls(return_wide = TRUE)
}
\arguments{
\item{return_wide}{}
}
\description{
Get ADV Summary Filing Data URLs for All Applicable Periods
}

