% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adv_functions.R
\name{get_data_all_cik_codes}
\alias{get_data_all_cik_codes}
\title{Get data all CIK code data}
\usage{
get_data_all_cik_codes(return_message = T)
}
\arguments{
\item{return_message}{}
}
\description{
Get data all CIK code data
}

