% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venture_data_functions.R
\name{get_data_all_ycombinator_graduates}
\alias{get_data_all_ycombinator_graduates}
\title{Get YCombinator Graduate Data}
\usage{
get_data_all_ycombinator_graduates(return_message = T)
}
\arguments{
\item{return_message}{}
}
\description{
Get YCombinator Graduate Data
}

