% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sec_functions.R
\name{get_data_cik_codes}
\alias{get_data_cik_codes}
\title{Get all data for all CIK registered entities}
\usage{
get_data_cik_codes(return_message = TRUE)
}
\arguments{
\item{return_message}{Return a message}
}
\description{
Get all data for all CIK registered entities
}

