% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crisp.R
\name{get_data_crsp_files}
\alias{get_data_crsp_files}
\title{CRSP file dictionary}
\usage{
get_data_crsp_files(include_summary_files = TRUE)
}
\arguments{
\item{include_summary_file}{\code{TRUE} (default) or \code{FALSE} to exclude}
}
\value{
a \code{data_frame}
}
\description{
This function returns information about available
CRSP index data
}
\examples{
get_data_crsp_files()
}
\references{
\href{http://crsp.com}{The Center for Research in Security Prices}
}
\seealso{
Other CRSP: \code{\link{get_data_crsp_indicies_constituents}},
  \code{\link{get_data_crsp_indicies_returns}}
}

