% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crisp.R
\name{get_data_crsp_indicies_constituents}
\alias{get_data_crsp_indicies_constituents}
\title{Get CRSP constituency data by index}
\usage{
get_data_crsp_indicies_constituents(nest_data = TRUE, return_message = TRUE)
}
\arguments{
\item{nest_data}{return nested data frame \code{TRUE, FALSE}}

\item{return_message}{return a message after data is parsed \code{TRUE, FALSE}}
}
\description{
Get CRSP constituency data by index
}
\examples{
get_data_crsp_indicies_constituents(nest_data = FALSE, return_message = TRUE)
}

