% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crisp.R
\name{get_data_crsp_indicies_returns}
\alias{get_data_crsp_indicies_returns}
\title{Get CRSP index return by month}
\usage{
get_data_crsp_indicies_returns(return_wide = FALSE, nest_data = TRUE,
  return_message = TRUE)
}
\arguments{
\item{return_wide}{return data in wide form \code{TRUE, FALSE}}

\item{nest_data}{return nested data frame \code{TRUE, FALSE}}

\item{return_message}{return a message after data is parsed \code{TRUE, FALSE}}
}
\description{
Get CRSP index return by month
}
\examples{
get_data_crsp_indicies_returns(return_wide = FALSE, nest_data = FALSE, return_message = TRUE)
}

