% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtcc.R
\name{get_data_dtcc_today}
\alias{get_data_dtcc_today}
\title{Get days cleared DTCC trades by asset type}
\usage{
get_data_dtcc_today(assets = NULL, nest_data = TRUE,
  return_message = TRUE)
}
\arguments{
\item{assets}{type of DTCC cleared financial product \code{NULL, COMMODITIES, CREDITS, EQUITIES, FOREX, RATES}}

\item{nest_data}{return a nested data frame \code{TRUE, FALSE}}

\item{return_message}{return a message \code{TRUE, FALSE}}
}
\description{
Get days cleared DTCC trades by asset type
}
\examples{
get_data_dtcc_today(assets = NULL, return_message = TRUE, nest_data = FALSE)
}

