% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtcc.R
\name{get_data_dtcc_trades}
\alias{get_data_dtcc_trades}
\title{DTCC trades by date and product}
\usage{
get_data_dtcc_trades(assets = NULL, include_today = TRUE,
  start_date = NULL, end_date = NULL, nest_data = TRUE,
  return_message = TRUE)
}
\arguments{
\item{assets}{type of DTCC cleared financial product \itemize{
\item \code{NULL}: returns all product types (default)
\item \code{COMMODITIES}: Commodities
\item \code{CREDITS}: Credit Default Swaps
\item \code{EQUITIES}: Equities
\item \code{FOREX}: Foreign Exchange
\item \code{RATES}: Interest Rates
}}

\item{include_today}{include today's trades}

\item{start_date}{date starting, must be in year-day-format}

\item{end_date}{date ending, must be in year-month-day format}

\item{nest_data}{\code{TRUE} return nested data frame}

\item{return_message}{\code{TRUE} return a message after data import}
}
\value{
where \code{nest_data} is \code{TRUE} a nested data_frame by asset and action
where \code{nest_data} is \code{FALSE} a data_frame

nested \code{data_frame} or \code{data_frame} if \code{nest_data = FALSE}

where \code{nest_data} is \code{TRUE} a nested data_frame by asset,
where \code{nest_data} is \code{FALSE} a data_frame
}
\description{
This function returns information on derivatives trades cleared by the
The Depository Trust & Clearing Corporation [DTCC] for specified dates
and product types.
}
\note{
Use \code{\link{get_data_dtcc_most_recent_trades}} for most recent trades
}
\examples{
\dontrun{
get_data_dtcc_trades()
get_data_dtcc_trades(start_date = "2017-01-16")
get_data_dtcc_trades(assets = c('credits', 'equities'), include_today = TRUE, start_date = '2017-01-10', end_date = Sys.Date(), nest_data = FALSE)
}
}
\references{
\href{http://dtcc.com}{The Depository Trust & Clearing Corporation}

\href{http://dtcc.com}{The Depository Trust & Clearing Corporation}
}
\seealso{
Other DTCC: \code{\link{get_data_dtcc_most_recent_trades}}

Other transaction data: \code{\link{get_data_dtcc_most_recent_trades}},
  \code{\link{get_data_nareit_mergers_acquisitions}},
  \code{\link{get_data_sec_securities_metrics_by_exchange}}
}
