% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgar_rf.R
\name{get_data_edgar_filing_streams}
\alias{get_data_edgar_filing_streams}
\title{SEC filing streams}
\usage{
get_data_edgar_filing_streams(start_date = "2017-02-15",
  end_date = Sys.Date(), only_most_recent_data = FALSE,
  index_type = "master", table_name_initial = "Filing Logs",
  parse_all_filings = FALSE, parse_complete_text_filings = FALSE,
  parse_form_d = FALSE, parse_13F = FALSE, parse_small_offerings = FALSE,
  parse_form_3_4s = FALSE, parse_asset_files = FALSE, parse_xbrl = FALSE,
  assign_to_environment = TRUE, nest_data = TRUE, return_message = TRUE)
}
\arguments{
\item{start_date}{starting date in year-month-date form}

\item{end_date}{ending date starting in year-month-date form}

\item{only_most_recent_data}{\code{TRUE} return only most recent day's filing stream}

\item{index_type}{type of index to parse \itemize{
\item \code{master}: parses master log (default)
\item \code{compamy}: parses company log
\item \code{filer}: parses filer log
}}

\item{nest_data}{\code{TRUE} return nested data frame}

\item{return_message}{\code{TRUE} return a message after data import}
}
\value{
nested \code{data_frame} or \code{data_frame} if \code{nest_data = FALSE}
}
\description{
This function parses daily
SEC filing log starting in 1994
for specified periods
}
\examples{
\dontrun{
get_data_edgar_filing_streams(start_date = "2016-01-01",
end_date = Sys.Date(), only_most_recent_data = FALSE, index_type = 'master',
nest_data = TRUE,
return_message = TRUE)
}
}
\references{
\href{http://sec.gov}{The Securities and Exchange Commission}
}
\seealso{
Other SEC: \code{\link{get_data_mmf_owned_debt_securities}},
  \code{\link{get_data_recent_insider_trades}},
  \code{\link{get_data_rf_leis}},
  \code{\link{get_data_rf_sec_13F_companies}},
  \code{\link{get_data_sec_bankruptcies}},
  \code{\link{get_data_sec_broker_dealers}},
  \code{\link{get_data_sec_ciks}},
  \code{\link{get_data_sec_closed_end_funds}},
  \code{\link{get_data_sec_cusips}},
  \code{\link{get_data_sec_failed_to_deliver_securities}},
  \code{\link{get_data_sec_filer}},
  \code{\link{get_data_sec_filing_entities}},
  \code{\link{get_data_sec_foia_requests}},
  \code{\link{get_data_sec_investment_companies}},
  \code{\link{get_data_sec_money_market_funds}},
  \code{\link{get_data_sec_municipal_advisors}},
  \code{\link{get_data_sec_securities_filing_counts}},
  \code{\link{get_data_sec_securities_metrics_by_exchange}},
  \code{\link{get_data_securities_offerings}},
  \code{\link{get_data_us_public_companies}},
  \code{\link{get_dictionary_sec_filing_codes}},
  \code{\link{get_dictionary_sec_form_codes}},
  \code{\link{get_dictionary_sic_codes}}
}
\concept{SEC}
\concept{filing search}
