% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgar_rf.R
\name{get_data_edgar_filings_most_recent}
\alias{get_data_edgar_filings_most_recent}
\title{Most recent EDGAR filings by typ}
\usage{
get_data_edgar_filings_most_recent(forms = c("All", "10-D", "10-K"),
  table_name_initial = "Recent Filings", parse_all_filings = TRUE,
  parse_form_d = FALSE, parse_complete_text_filings = FALSE,
  parse_13F = FALSE, parse_small_offerings = FALSE,
  parse_form_3_4s = FALSE, parse_asset_files = FALSE, parse_xbrl = FALSE,
  assign_to_environment = TRUE, nest_data = FALSE, return_message = TRUE)
}
\arguments{
\item{return_message}{}
}
\description{
Most recent EDGAR filings by typ
}
