% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adv_functions.R
\name{get_data_finra_entities}
\alias{get_data_finra_entities}
\title{Get FINRA registered firms and associated metadata}
\usage{
get_data_finra_entities(search_names = c("Thrive", "Rockwood Capital", "EJF",
  "Simple Capital"), ocr_pdf = TRUE, return_message = TRUE)
}
\arguments{
\item{search_names}{vector of names to search}

\item{return_message}{return a message upon parsing \code{TRUE, FALS}}
}
\description{
Get FINRA registered firms and associated metadata
}
\examples{
get_data_finra_entities(search_name = c('EJF Capital', 'Goldman Sachs'), ocr_pdf = TRUE)
}

