% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_and_filed_functions.R
\name{get_data_form_ds}
\alias{get_data_form_ds}
\title{Get Form D Filing Data}
\usage{
get_data_form_ds(industries = NULL, years_filings = NULL,
  return_message = TRUE)
}
\description{
Get Form D Filing Data
}

