% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debt_functions.R
\name{get_data_fred_index_symbol_time_series}
\alias{get_data_fred_index_symbol_time_series}
\title{Get time series data for a given FRED symbol}
\usage{
get_data_fred_index_symbol_time_series(symbol = "DGS10", return_wide = F)
}
\arguments{
\item{return_wide}{}
}
\description{
Get time series data for a given FRED symbol
}
\examples{
get_data_fred_index_symbol_time_series(symbol = 'DGS10',return_wide = F)
}

