% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fred_api.R
\name{get_data_fred_tags}
\alias{get_data_fred_tags}
\title{FRED tag search}
\usage{
get_data_fred_tags(tags = NULL, return_message = T, nest_data = F)
}
\arguments{
\item{tags}{vector of tags}

\item{return_message}{if \code{TRUE} returns message}

\item{nest_data}{if \code{TRUE} nests data}
}
\value{
\code{data_frame}
}
\description{
FRED tag search
}
\examples{
get_data_fred_tags(tags = c("spread", "swaps"))
}
