% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fred_api.R
\name{get_data_fred_terms_ids}
\alias{get_data_fred_terms_ids}
\title{FRED Terms Series IDs}
\usage{
get_data_fred_terms_ids(search_terms = c("China", "Property"),
  nest_data = FALSE, return_message = TRUE)
}
\arguments{
\item{nest_data}{}
}
\description{
This function returns
matching FRED series ID for a given
search term.  The maximum number of results is
100 for broader search use #' \code{\link{get_dictionary_all_fred_series_ids}} function.
}
