% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debt_functions.R
\name{get_data_futures_indicies}
\alias{get_data_futures_indicies}
\title{Get futures data for specified index}
\usage{
get_data_futures_indicies(future_indicies = c("30 Day Fed Funds",
  "Eurodollars", "1 Month Libor", "Euribor"), return_wide = F)
}
\arguments{
\item{return_wide}{}
}
\description{
Get futures data for specified index
}

