% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debt_functions.R
\name{get_data_index_symbol}
\alias{get_data_index_symbol}
\title{Get FRED Index Data}
\usage{
get_data_index_symbol(symbol = "DGS10", return_wide = F)
}
\arguments{
\item{return_wide}{}
}
\description{
Get FRED Index Data
}

