% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debt_functions.R
\name{get_data_index_symbol_current}
\alias{get_data_index_symbol_current}
\title{Title}
\usage{
get_data_index_symbol_current(symbol = "^TNX", return_wide = F)
}
\arguments{
\item{return_wide}{}
}
\description{
Title
}

