% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_and_filed_functions.R
\name{get_data_insider_trades_most_recent}
\alias{get_data_insider_trades_most_recent}
\title{Get data money market owned securities}
\usage{
get_data_insider_trades_most_recent()
}
\arguments{
\item{return_message}{}
}
\description{
Get data money market owned securities
}

