% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_and_filed_functions.R
\name{get_data_leis}
\alias{get_data_leis}
\title{Get US domiciled LEI issuers}
\usage{
get_data_leis(return_message = TRUE)
}
\arguments{
\item{return_message}{}
}
\description{
Get US domiciled LEI issuers
}

