% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debt_functions.R
\name{get_data_libor_current}
\alias{get_data_libor_current}
\title{Get current LIBOR data}
\usage{
get_data_libor_current(return_wide = T)
}
\arguments{
\item{return_wide}{}
}
\description{
Get current LIBOR data
}

