% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_and_filed_functions.R
\name{get_data_location_codes}
\alias{get_data_location_codes}
\title{Get location code data}
\usage{
get_data_location_codes()
}
\description{
Get location code data
}
\examples{
get_data_location_codes()
}

