% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adv_functions.R
\name{get_data_managers_adv_filings}
\alias{get_data_managers_adv_filings}
\title{Get detailed ADV form data for specified search name or CRD id}
\usage{
get_data_managers_adv_filings(search_names = c("Blackstone Real Estate"),
  crd_ids = NULL, all_sections = T, section_names = c("Registration",
  "Identifying Information", "Organization", "Successions",
  "Private Fund Reporting", "Direct Manager Owners", "Indirect Manager Owners",
  "Other Manager Information", "Manager Signatories"), flatten_tables = T,
  gather_data = F, assign_to_enviornment = T)
}
\arguments{
\item{assign_to_enviornment}{}
}
\description{
Get detailed ADV form data for specified search name or CRD id
}

