% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trading_view.R
\name{get_data_market_events}
\alias{get_data_market_events}
\title{Get Market Events}
\usage{
get_data_market_events(return_message = TRUE)
}
\arguments{
\item{return_message}{}
}
\description{
Returns a list of global financial
and economics events
}
\examples{
get_data_market_events(retur_message = TRUE)
}
