% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_and_filed_functions.R
\name{get_data_mmf_owned_debt_securities}
\alias{get_data_mmf_owned_debt_securities}
\title{Get data money market owned securities}
\usage{
get_data_mmf_owned_debt_securities(return_message = TRUE)
}
\arguments{
\item{return_message}{}
}
\description{
Get data money market owned securities
}

