% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicies.R
\name{get_data_msci_indicies}
\alias{get_data_msci_indicies}
\title{Get MSCI indicies}
\usage{
get_data_msci_indicies()
}
\value{
A data_frame with all MSCI indicies
}
\description{
Get MSCI indicies
}
\examples{
get_data_msci_indicies()
}

