% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicies.R
\name{get_data_msci_indicies_constituents}
\alias{get_data_msci_indicies_constituents}
\title{Get MSCI indicies constituents}
\usage{
get_data_msci_indicies_constituents(indicies = NULL, nest_data = TRUE,
  return_message = TRUE)
}
\arguments{
\item{indicies}{specific indicies, \code{NULL} returns all}

\item{nest_data}{return a nested data frame \code{TRUE, FALSE}}

\item{return_message}{return a message \code{TRUE, FALSE}}
}
\description{
Get MSCI indicies constituents
}
\examples{
get_data_msci_indicies_constituents(indicies = "NORTH AMERICA", nest_data = FALSE, return_message = TRUE)

get_data_msci_indicies_constituents(indicies = NULL, nest_data = TRUE, return_message = TRUE)
}

