% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nareit.R
\name{get_data_nareit_annual_subsector_returns}
\alias{get_data_nareit_annual_subsector_returns}
\title{Get NAREIT subsector returns}
\usage{
get_data_nareit_annual_subsector_returns(return_wide = TRUE,
  return_message = TRUE)
}
\arguments{
\item{return_message}{}
}
\description{
Get NAREIT subsector returns
}
\examples{
get_data_nareit_annual_subsector_returns(return_wide = TRUE, return_message = TRUE)
}

