% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nareit.R
\name{get_data_nareit_capital_raises}
\alias{get_data_nareit_capital_raises}
\title{Get NAREIT capital raise data by year and form of capital}
\usage{
get_data_nareit_capital_raises(capital_type = NULL, nest_data = TRUE,
  return_message = TRUE)
}
\arguments{
\item{capital_type}{form of capital raised \code{NULL, IPO, SECONDARY, DEBT}}

\item{nest_data}{return a nested data frame \code{TRUE, FALSE}}

\item{return_message}{return a message \code{TRUE, FALSE}}
}
\description{
Get NAREIT capital raise data by year and form of capital
}
\examples{
get_data_nareit_capital_raises(capital_type = NULL, nest_data = FALSE,return_message = TRUE)
}

