% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nareit.R
\name{get_data_nareit_entities}
\alias{get_data_nareit_entities}
\title{Get NAREIT constituent data}
\usage{
get_data_nareit_entities(include_private_non_public = TRUE,
  return_message = TRUE)
}
\arguments{
\item{include_private_non_public}{include privately traded or over-the-counter traded REITs \code{TRUE, FALSE}}

\item{return_message}{return a message \code{TRUE, FALSE}}
}
\description{
Get NAREIT constituent data
}
\examples{
get_data_nareit_entities(include_private_non_public = TRUE, return_message = TRUE)
}

