% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nareit.R
\name{get_data_nareit_historic_monthly_returns}
\alias{get_data_nareit_historic_monthly_returns}
\title{Get monthly historic NAREIT returns by sector}
\usage{
get_data_nareit_historic_monthly_returns(return_wide = TRUE,
  return_message = TRUE)
}
\arguments{
\item{return_wide}{return data in wide form \code{TRUE, FALSE}}

\item{return_message}{return a message \code{T}}
}
\description{
Get monthly historic NAREIT returns by sector
}
\examples{
get_data_nareit_historic_monthly_returns(return_wide = FALSE)
}

