% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nareit.R
\name{get_data_nareit_industry_tracker}
\alias{get_data_nareit_industry_tracker}
\title{NAREIT quarterly industry tracking data}
\usage{
get_data_nareit_industry_tracker(nest_data = FALSE, return_wide = TRUE,
  return_message = TRUE)
}
\arguments{
\item{nest_data}{\code{TRUE} return nested data frame}

\item{return_wide}{\code{TRUE} return data in wide form}

\item{return_message}{\code{TRUE} return a message after data import}
}
\value{
nested \code{data_frame} or \code{data_frame} if \code{nest_data = FALSE}
}
\description{
This function acquires data on NAREIT
performance metrics including Funds From Operation [FFO],
Dividends, and Net Operating Income from 2000.
}
\examples{
\dontrun{
get_data_nareit_industry_tracker(return_message = TRUE, return_wide = TRUE)
}
}
\seealso{
Other NAREIT: \code{\link{get_data_nareit_annual_subsector_returns}},
  \code{\link{get_data_nareit_capital_raises}},
  \code{\link{get_data_nareit_constituent_years}},
  \code{\link{get_data_nareit_entities}},
  \code{\link{get_data_nareit_mergers_acquisitions}},
  \code{\link{get_data_nareit_monthly_returns}},
  \code{\link{get_data_nareit_notable_properties}},
  \code{\link{get_data_nareit_property_msa}},
  \code{\link{get_data_nareit_state_info}},
  \code{\link{get_data_reit_funds}}
}
\concept{NAREIT}
