% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nareit.R
\name{get_data_nareit_notable_properties}
\alias{get_data_nareit_notable_properties}
\title{NAREIT notable properties}
\usage{
get_data_nareit_notable_properties()
}
\value{
\code{data_frame}
}
\description{
This function returns information about notable NAREIT identified properties.
}
\examples{
get_data_nareit_notable_properties()
}
\references{
\href{http://nareit.org}{National Association of Real Estate Investment Trusts}
}
\seealso{
Other NAREIT: \code{\link{get_data_nareit_annual_subsector_returns}},
  \code{\link{get_data_nareit_capital_raises}},
  \code{\link{get_data_nareit_constituent_years}},
  \code{\link{get_data_nareit_entities}},
  \code{\link{get_data_nareit_industry_tracker}},
  \code{\link{get_data_nareit_mergers_acquisitions}},
  \code{\link{get_data_nareit_monthly_returns}},
  \code{\link{get_data_nareit_property_msa}},
  \code{\link{get_data_nareit_state_info}},
  \code{\link{get_data_reit_funds}},
  \code{\link{get_reit_entity_dictionary}}

Other property data: \code{\link{get_data_nareit_property_msa}},
  \code{\link{get_data_nareit_state_info}}
}
