% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nareit.R
\name{get_data_nareit_state_info}
\alias{get_data_nareit_state_info}
\title{Get NAREIT member property holdings by state}
\usage{
get_data_nareit_state_info(include_reit_hq = TRUE,
  include_reit_holdings = TRUE, nest_data = TRUE, return_message = TRUE)
}
\arguments{
\item{include_reit_hq}{include REIT headquarter locations \code{TRUE, FALSE}}

\item{include_reit_holdings}{include REIT aggregate holdings \code{TRUE, FALSE}}

\item{nest_data}{return a nested data frame \code{TRUE, FALSE}}

\item{return_message}{return a message \code{TRUE, FALSE}}
}
\description{
Get NAREIT member property holdings by state
}
\examples{
get_data_nareit_state_info(include_reit_hq = TRUE, include_reit_holdings = TRUE, nest_data = TRUE, return_message = TRUE)
}

