% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterfall_functions.R
\name{get_data_promote_structure}
\alias{get_data_promote_structure}
\title{Parse Vector of Promotes into a data frame}
\usage{
get_data_promote_structure(promote_structures = c("20 over a 12", "30 / 18",
  "40 over a 10x"), return_wide = F)
}
\arguments{
\item{promote_structures}{A vector of promote structures}

\item{return_wide}{}
}
\description{
Parse Vector of Promotes into a data frame
}
\examples{
get_data_promote_structure(promote_structures = c("20 over a 12", '30 / 18', "40 over a 10x"), return_wide = T)
}

