% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_and_filed_functions.R
\name{get_data_recent_insider_trades}
\alias{get_data_recent_insider_trades}
\title{Get most recent insider trades}
\usage{
get_data_recent_insider_trades(return_message = TRUE)
}
\description{
Get most recent insider trades
}

