% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_and_filed_functions.R
\name{get_data_sec_13F_companies}
\alias{get_data_sec_13F_companies}
\title{Get 13F owned compaines}
\usage{
get_data_sec_13F_companies(return_message = TRUE)
}
\arguments{
\item{return_message}{}
}
\description{
Get 13F owned compaines
}

