% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adv_functions.R
\name{get_data_sec_adv_manager_sitemap}
\alias{get_data_sec_adv_manager_sitemap}
\title{Returns data frame of SEC ADV sitemap}
\usage{
get_data_sec_adv_manager_sitemap()
}
\description{
Returns data frame of SEC ADV sitemap
}
