% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sec_functions.R
\name{get_data_sec_bankruptcies}
\alias{get_data_sec_bankruptcies}
\title{Get SEC registered public bankruptcies}
\usage{
get_data_sec_bankruptcies(nest_data = FALSE, return_message = TRUE)
}
\arguments{
\item{nest_data}{return a nested data frame}

\item{return_message}{return a message}
}
\description{
Get SEC registered public bankruptcies
}
\examples{
get_data_sec_bankruptcies(nest_data = FALSE, return_message = TRUE)
}

