% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sec_functions.R
\name{get_data_sec_cik_issuers}
\alias{get_data_sec_cik_issuers}
\title{Returns all SEC CIK registered entities}
\usage{
get_data_sec_cik_issuers(return_message = TRUE)
}
\arguments{
\item{return_message}{return a message after parsing data}
}
\description{
Returns all SEC CIK registered entities
}
\examples{
get_data_sec_cik_issuers(return_message = TRUE)
}

