% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sec_functions.R
\name{get_data_sec_closed_end_funds}
\alias{get_data_sec_closed_end_funds}
\title{Get SEC registered closed end funds}
\usage{
get_data_sec_closed_end_funds(only_most_recent = FALSE, years = 2016,
  nest_data = TRUE, return_message = TRUE)
}
\arguments{
\item{only_most_recent}{return only the most recent year}

\item{years}{vector of years to search starting in 2012}

\item{nest_data}{return a nested data frame}

\item{return_message}{return a message}
}
\description{
Get SEC registered closed end funds
}
\examples{
get_data_sec_closed_end_funds()
}

