% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sec_functions.R
\name{get_data_sec_failed_to_deliver_securities}
\alias{get_data_sec_failed_to_deliver_securities}
\title{Get SEC Failed to Deliver Securities Data}
\usage{
get_data_sec_failed_to_deliver_securities(only_most_recent = TRUE,
  years = 2016, months = NULL, nest_data = TRUE, return_message = TRUE)
}
\arguments{
\item{return_message}{}
}
\description{
Get SEC Failed to Deliver Securities Data
}
\examples{
get_data_sec_failed_to_deliver_securities(nest_data = )
}

