% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_and_filed_functions.R
\name{get_data_sec_filer}
\alias{get_data_sec_filer}
\title{Get SEC filer data}
\usage{
get_data_sec_filer(entity_names = NULL, tickers = NULL, ciks = NULL,
  tables = NULL, nest_data = TRUE, include_subsidiaries = TRUE,
  include_13Fs = TRUE, include_asset_data = TRUE,
  include_small_offering_data = TRUE, assign_to_environment = TRUE,
  return_message = TRUE)
}
\description{
Get SEC filer data
}

