% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_and_filed_functions.R
\name{get_data_sec_filing_entities}
\alias{get_data_sec_filing_entities}
\title{Get SEC Filing Entities for Search Names}
\usage{
get_data_sec_filing_entities(entity_names = c("Rockwood Capital", "Vornado",
  "Two Sigma"), return_message = TRUE)
}
\arguments{
\item{return_message}{}
}
\description{
Get SEC Filing Entities for Search Names
}

