% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_and_filed_functions.R
\name{get_data_sec_filing_streams}
\alias{get_data_sec_filing_streams}
\title{Get SEC Filing Stream Data}
\usage{
get_data_sec_filing_streams(filers = c("All", "Corporate Insider",
  "Companies", "Investment Company"), filing_names = c("All",
  "Annual Reports", "Quarterly Reports", "Current Reports", "Other Reports",
  "Registrations", "Private Offerings", "Ownership", "Prospectuses",
  "Exemptions", "Withdrawals", "Correspondence", "Proxy Statements",
  "Confidential"), return_message = TRUE)
}
\description{
Get SEC Filing Stream Data
}

