% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_and_filed_functions.R
\name{get_data_sec_form_ds}
\alias{get_data_sec_form_ds}
\title{Get SEC Form D data by industry}
\usage{
get_data_sec_form_ds(industries = NULL, form_years = NULL, months = NULL,
  return_message = TRUE)
}
\description{
Get SEC Form D data by industry
}

