% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sec_functions.R
\name{get_data_sec_investment_companies}
\alias{get_data_sec_investment_companies}
\title{Get SEC registered investment companies}
\usage{
get_data_sec_investment_companies(nest_data = TRUE)
}
\description{
Get SEC registered investment companies
}

