% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sec_functions.R
\name{get_data_sec_money_market_funds}
\alias{get_data_sec_money_market_funds}
\title{Get SEC registered money market funds}
\usage{
get_data_sec_money_market_funds(only_most_recent = TRUE, years = NULL,
  months = NULL, nest_data = TRUE, return_message = TRUE)
}
\arguments{
\item{only_most_recent}{return only most recent year}

\item{years}{years to include}

\item{months}{months to include}

\item{return_message}{return a message}
}
\description{
Get SEC registered money market funds
}
\examples{
get_data_sec_money_market_funds(only_most_recent = TRUE, nest_data = FALSE)
get_data_sec_money_market_funds(only_most_recent = FALSE, years = 2016)
}

