% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sec_functions.R
\name{get_data_sec_municipal_advisors}
\alias{get_data_sec_municipal_advisors}
\title{Get SEC registered municipal dealers}
\usage{
get_data_sec_municipal_advisors(only_most_recent = TRUE, years = NULL,
  months = NULL, nest_data = FALSE, return_message = TRUE)
}
\arguments{
\item{only_most_recent}{return only most recent year}

\item{years}{years to include}

\item{months}{months to include}

\item{nest_data}{nest the data}

\item{return_message}{return a message}
}
\description{
Get SEC registered municipal dealers
}
\examples{
get_data_sec_municipal_advisors(only_most_recent = TRUE)
get_data_sec_municipal_advisors(only_most_recent = FALSE, years = 2016:2017, nest_data = TRUE)
}

