% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_and_filed_functions.R
\name{get_data_sec_mutual_funds}
\alias{get_data_sec_mutual_funds}
\title{Get SEC registered mutual funds}
\usage{
get_data_sec_mutual_funds(return_message = TRUE)
}
\description{
Get SEC registered mutual funds
}

