% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_and_filed_functions.R
\name{get_data_sec_private_filers}
\alias{get_data_sec_private_filers}
\title{Get SEC Private Filer data}
\usage{
get_data_sec_private_filers(entity_names = NULL, ciks = NULL,
  tables = NULL, assign_to_environment = TRUE, return_message = TRUE)
}
\description{
Get SEC Private Filer data
}

