% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_and_filed_functions.R
\name{get_data_sec_rules}
\alias{get_data_sec_rules}
\title{SEC Prefix Laws}
\usage{
get_data_sec_rules()
}
\value{
a  \code{data_frame}
}
\description{
This function returns prefix
descriptions of SEC rules and the
law that created the rule
}
\seealso{
Other dictionary: \code{\link{get_data_location_codes}},
  \code{\link{get_data_sic_codes}},
  \code{\link{get_dictionary_sec_filing_codes}},
  \code{\link{get_dictionary_sec_form_codes}},
  \code{\link{get_reit_entity_dictionary}}
}

