% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_and_filed_functions.R
\name{get_data_sec_securities_filing_counts}
\alias{get_data_sec_securities_filing_counts}
\title{Get securities registration filing counts}
\usage{
get_data_sec_securities_filing_counts(return_message = TRUE)
}
\description{
Get securities registration filing counts
}

