% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sec_functions.R
\name{get_data_sec_securities_metrics_by_exchange}
\alias{get_data_sec_securities_metrics_by_exchange}
\title{Get SEC aggregate securities metrics}
\usage{
get_data_sec_securities_metrics_by_exchange(only_most_recent = FALSE,
  years = 2016, months = NULL, nest_data = FALSE, return_message = TRUE)
}
\arguments{
\item{only_most_recent}{return only most recent year}

\item{years}{years to include}

\item{months}{months to include}

\item{nest_data}{nest the data}

\item{return_message}{return a message}
}
\description{
Get SEC aggregate securities metrics
}
\examples{
get_data_sec_securities_metrics_by_exchange()
}

