% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgar_rf.R
\name{get_data_sec_tickers_info}
\alias{get_data_sec_tickers_info}
\title{Get Ticker SEC company information}
\usage{
get_data_sec_tickers_info(tickers = c("VNO", "NVDA", "FB"),
  return_message = TRUE)
}
\arguments{
\item{tickers}{character vector of ticker symbols}

\item{return_message}{if \code{true} return a message}
}
\value{
a \code{data_frame}
}
\description{
Get Ticker SEC company information
}
\examples{
get_data_sec_tickers_info(tickers = c("BXP", "AVB", "AAPL"))
}
