% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_and_filed_functions.R
\name{get_data_securities_filing_counts}
\alias{get_data_securities_filing_counts}
\title{Get Securities Filing Counts}
\usage{
get_data_securities_filing_counts(return_message = TRUE)
}
\description{
Get Securities Filing Counts
}

