% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgar_rf.R
\name{get_data_securities_offerings}
\alias{get_data_securities_offerings}
\title{Securities offerings}
\usage{
get_data_securities_offerings(year_forms = NULL, forms = NULL,
  nest_data = FALSE, return_message = TRUE)
}
\arguments{
\item{year_forms}{years to search starting in 1994, \code{NULL} returns all years}

\item{forms}{forms to use in search \itemize{
\item \code{NULL}: returns all (default)
\item \code{D}: Exempt offerings
\item \code{W}: Secondary sales
\item \code{S-1}: Initial public offerings
\item \code{S-4}: Initial public offering withdrawls
}}

\item{nest_data}{\code{TRUE} return nested data frame}

\item{return_message}{\code{TRUE} return a message after data import}
}
\value{
where \code{nest_data} is \code{TRUE} a nested data_frame by asset,
where \code{nest_data} is \code{FALSE} a data_frame
}
\description{
This function returns securities offering
data by form type and year from 1994 onward.
}
\examples{
#' \dontrun{

## All Securities Filings
get_data_securities_offerings(year_forms = NULL, forms = NULL, return_message = TRUE, nest_data = TRUE)

## IPOs since 1999
get_data_securities_offerings(year_forms = 1999:2017, forms = "S-1", return_message = TRUE, nest_data = FALSE)
}
}
\seealso{
Other SEC: \code{\link{get_data_edgar_filing_streams}},
  \code{\link{get_data_mmf_owned_debt_securities}},
  \code{\link{get_data_recent_insider_trades}},
  \code{\link{get_data_rf_leis}},
  \code{\link{get_data_rf_sec_13F_companies}},
  \code{\link{get_data_sec_bankruptcies}},
  \code{\link{get_data_sec_broker_dealers}},
  \code{\link{get_data_sec_ciks}},
  \code{\link{get_data_sec_closed_end_funds}},
  \code{\link{get_data_sec_cusips}},
  \code{\link{get_data_sec_failed_to_deliver_securities}},
  \code{\link{get_data_sec_filer}},
  \code{\link{get_data_sec_filing_entities}},
  \code{\link{get_data_sec_foia_requests}},
  \code{\link{get_data_sec_investment_companies}},
  \code{\link{get_data_sec_money_market_funds}},
  \code{\link{get_data_sec_municipal_advisors}},
  \code{\link{get_data_sec_securities_filing_counts}},
  \code{\link{get_data_sec_securities_metrics_by_exchange}},
  \code{\link{get_data_sec_xbrl_periods}},
  \code{\link{get_data_us_public_companies}},
  \code{\link{get_dictionary_sec_filing_codes}},
  \code{\link{get_dictionary_sec_form_codes}},
  \code{\link{get_dictionary_sic_codes}}

Other entity search: \code{\link{get_data_adv_managers_current_period_summary}},
  \code{\link{get_data_adv_managers_filings}},
  \code{\link{get_data_adv_managers_metadata}},
  \code{\link{get_data_adv_managers_periods_summaries}},
  \code{\link{get_data_finra_entities}},
  \code{\link{get_data_nareit_entities}},
  \code{\link{get_data_reit_funds}},
  \code{\link{get_data_rf_leis}},
  \code{\link{get_data_rf_sec_13F_companies}},
  \code{\link{get_data_sec_bankruptcies}},
  \code{\link{get_data_sec_broker_dealers}},
  \code{\link{get_data_sec_ciks}},
  \code{\link{get_data_sec_closed_end_funds}},
  \code{\link{get_data_sec_cusips}},
  \code{\link{get_data_sec_filer}},
  \code{\link{get_data_sec_filing_entities}},
  \code{\link{get_data_sec_investment_companies}},
  \code{\link{get_data_sec_money_market_funds}},
  \code{\link{get_data_sec_municipal_advisors}},
  \code{\link{get_data_sec_xbrl_periods}},
  \code{\link{get_data_us_public_companies}},
  \code{\link{get_data_ycombinator_alumni}}

Other securities search: \code{\link{get_data_mmf_owned_debt_securities}},
  \code{\link{get_data_rf_sec_13F_companies}},
  \code{\link{get_data_sec_cusips}}

Other Rank and Filed: \code{\link{get_data_mmf_owned_debt_securities}},
  \code{\link{get_data_recent_insider_trades}},
  \code{\link{get_data_rf_leis}},
  \code{\link{get_data_rf_sec_13F_companies}},
  \code{\link{get_data_sec_filer}},
  \code{\link{get_data_sec_filing_entities}},
  \code{\link{get_data_us_public_companies}}
}
\concept{Rank and Filed}
\concept{SEC}
\concept{entity search}
\concept{securities search}
