% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgar_rf.R
\name{get_data_sic_naics_codes}
\alias{get_data_sic_naics_codes}
\title{SIC Codes}
\usage{
get_data_sic_naics_codes(filter_duplicates = TRUE)
}
\arguments{
\item{filter_duplicates}{\code{TRUE} removes duplicate entries}
}
\value{
a \code{data_frame}
}
\description{
This function returns Standard Industrial Classification [SIC]
codes and the corresponding
North American Industry Classification System [NAICS] identification.
}
\examples{
get_data_sic_naics_codes(filter_duplicates = TRUE)
}
\seealso{
Other dictionary: \code{\link{get_data_location_codes}},
  \code{\link{get_data_sec_rules}},
  \code{\link{get_dictionary_all_fred_series_ids}},
  \code{\link{get_dictionary_sec_filing_codes}},
  \code{\link{get_dictionary_sec_form_codes}},
  \code{\link{get_dictionary_sic_codes}}
}
\concept{dictionary}
