% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finbox.R
\name{get_data_tickers_finbox}
\alias{get_data_tickers_finbox}
\title{Finbox ticker data}
\usage{
get_data_tickers_finbox(tickers = c("VNO", "BXP"),
  assign_to_environment = TRUE)
}
\arguments{
\item{tickers}{\code{vector} of tickers}

\item{assign_to_environment}{}
}
\description{
Finbox ticker data
}
