% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markit.R
\name{get_data_tickers_trades}
\alias{get_data_tickers_trades}
\title{Tickers trades}
\usage{
get_data_tickers_trades(tickers = c("PEI", "bxp"),
  return_message = TRUE)
}
\arguments{
\item{return_message}{}
}
\description{
Tickers trades
}
\examples{
get_data_tickers_trades(tickers = c("VNO", "BXP", "FB"))

}
