% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trading_view.R
\name{get_data_tickers_tradingview_news}
\alias{get_data_tickers_tradingview_news}
\title{Tradingview tickers news}
\usage{
get_data_tickers_tradingview_news(tickers = c("FB", "AAPL", "NFLX",
  "GOOG", "VNO", "EQR", "BXP"), return_message = TRUE,
  nest_data = FALSE)
}
\description{
Returns news data for specified tickers
}
\examples{
get_data_tickers_tradingview_news(tickers = c("VNO", "AVB", "PEI"), return_message = TRUE, nest_data = FALSE)
}
