% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venture_data_functions.R
\name{get_data_unicorn_valuations}
\alias{get_data_unicorn_valuations}
\title{Get Unicorn Valuation Data}
\usage{
get_data_unicorn_valuations(return_message = TRUE)
}
\description{
Get Unicorn Valuation Data
}

