% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inflation.R
\name{get_data_us_inflation}
\alias{get_data_us_inflation}
\title{US Inflation Adjustment Table}
\usage{
get_data_us_inflation(base_year = 1913)
}
\arguments{
\item{base_year}{1913 or 1800}
}
\value{
a \code{data_frame}
}
\description{
Gets a table to adjust prices to current dollars
}
\examples{
get_data_us_inflation(base_year = 1913)
}
