% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_and_filed_functions.R
\name{get_data_us_public_companies}
\alias{get_data_us_public_companies}
\title{Get United States public securities descriptive data}
\usage{
get_data_us_public_companies(merge_type = NULL, return_message = TRUE)
}
\arguments{
\item{return_message}{}
}
\description{
Get United States public securities descriptive data
}

