% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_and_filed_functions.R
\name{get_data_us_tickers}
\alias{get_data_us_tickers}
\title{Get ticker information for United States traded public entities}
\usage{
get_data_us_tickers(return_message = TRUE)
}
\arguments{
\item{return_message}{}
}
\description{
Get ticker information for United States traded public entities
}
\examples{
get_data_us_tickers(return_message = TRUE)
}

