% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venture_data_functions.R
\name{get_data_ycombinator_alumni}
\alias{get_data_ycombinator_alumni}
\title{Get data on Y-Combinator alumni companies}
\usage{
get_data_ycombinator_alumni(nest_data = FALSE, return_message = T)
}
\arguments{
\item{nest_data}{returns a nested data frame \code{TRUE, FALSE}}

\item{return_message}{return a message that includes a random YC company, \code{TRUE, FALSE}}
}
\description{
Get data on Y-Combinator alumni companies
}
\examples{
get_data_ycombinator_alumni()
}

