% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_and_filed_functions.R
\name{get_data_years_securities_offerings}
\alias{get_data_years_securities_offerings}
\title{Get Securities Offering data}
\usage{
get_data_years_securities_offerings(year_forms = NULL, forms = NULL,
  return_message = TRUE)
}
\arguments{
\item{return_message}{}
}
\description{
Get Securities Offering data
}

