% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finbox.R
\name{get_dictionary_finbox_companies}
\alias{get_dictionary_finbox_companies}
\title{FinboxIO Companies}
\usage{
get_dictionary_finbox_companies(return_message = TRUE)
}
\arguments{
\item{return_message}{if \code{TRUE} return a message}
}
\description{
Returns companies with FinBOXIO data
}
\examples{
get_dictionary_finbox_companies()
}
