% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_and_filed_functions.R
\name{get_dictionary_form_d_categories}
\alias{get_dictionary_form_d_categories}
\title{Get Form-D Category dictionry}
\usage{
get_dictionary_form_d_categories()
}
\description{
Get Form-D Category dictionry
}
\examples{
get_dictionary_form_d_categories()
}

