% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_and_filed_functions.R
\name{get_dictionary_sec_rules}
\alias{get_dictionary_sec_rules}
\title{SEC Rule dictionary}
\usage{
get_dictionary_sec_rules()
}
\description{
This function retuns a
dictionary of SEC rules
}
\examples{
get_dictionary_sec_rules()
}

