% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adv_functions.R
\name{get_manager_sec_adv_actual_url}
\alias{get_manager_sec_adv_actual_url}
\title{Title}
\usage{

  get_manager_sec_adv_actual_url(url = "http://www.adviserinfo.sec.gov/IAPD/crd_iapd_AdvVersionSelector.aspx?ORG_PK=146629")
}
\arguments{
\item{url}{}
}
\description{
Title
}

