% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debt_functions.R
\name{get_split_mixed_char_title}
\alias{get_split_mixed_char_title}
\title{Get Split Title}
\usage{
get_split_mixed_char_title(x = "namePersonActor")
}
\arguments{
\item{x}{}
}
\value{

}
\description{
Get Split Title
}

