% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fred_api.R
\name{import_rda_file}
\alias{import_rda_file}
\title{Import RDA file}
\usage{
import_rda_file(file = "Desktop/fred/fred_series_tags.rda")
}
\arguments{
\item{file}{a file path}
}
\description{
Import RDA file
}
