% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/real_estate_entities.R
\name{irei_entitites}
\alias{irei_entitites}
\title{IREI Entities}
\source{
IREI
}
\usage{
irei_entitites(filter_type = NULL, return_message = TRUE)
}
\arguments{
\item{filter_type}{filter entities types if \code{NULL} returns all options \itemize{
\item Accounting
\item Brokerage Firms
\item Construction Companies
\item Consultants
\item Developers and Operating Companies
\item Investment Bankers
\item Investment Managers
\item Law Firms
\item Real Estate Finance
\item Real Estate Recruiters
\item Real Estate Research & Information Providers
\item Seminar & Conference Producers
\item Tax-exempt Funds
\item Trade Associations & Organizations
}}

\item{return_message}{if \code{TRUE} returns a message}
}
\value{
a \code{tibble}
}
\description{
Acquires a list of IREI tracked entities
from 14 different parts of the "Commercial Real Estate Jungle"
}
\examples{
\dontrun{
irei_entitites(filter_type = NULL)
irei_entities(filter_type = c("Investment Managers", "Tax-exempt Funds"))
}
}
