% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgar_rf.R
\name{location_codes}
\alias{location_codes}
\title{Locations}
\usage{
location_codes()
}
\value{
a \code{tibble}
}
\description{
This function returns location codes
and the name the location
}
\examples{
location_codes()
}
\seealso{
Other dictionary: 
\code{\link{dictionary_fred_ids}()},
\code{\link{dictionary_sec_filing_codes}()},
\code{\link{dictionary_sec_form_codes}()},
\code{\link{dictionary_sic_codes}()},
\code{\link{sec_rules}()},
\code{\link{sic_naics_codes}()}
}
\concept{dictionary}
