% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trading_view.R
\name{market_events}
\alias{market_events}
\title{Get Market Events}
\usage{
market_events(return_message = TRUE)
}
\arguments{
\item{return_message}{}
}
\description{
Returns a list of global financial
and economics events
}
\examples{
market_events(return_message = TRUE)
}
