% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicies.R
\name{msci_indicies}
\alias{msci_indicies}
\title{MSCI indicies}
\usage{
msci_indicies()
}
\value{
\code{tibble}
}
\description{
Returns all MSCI indicies.  This function
can be used to find indicies to search with
\code{\link{msci_indicies_constituents}}
to specify indicies to extract constituents.
}
\examples{
msci_indicies()
}
\seealso{
Other MSCI: 
\code{\link{msci_indicies_constituents}()},
\code{\link{msci_realtime_index_values}()}
}
\concept{MSCI}
