% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicies.R
\name{msci_realtime_index_values}
\alias{msci_realtime_index_values}
\title{MSCI indicies values}
\usage{
msci_realtime_index_values(return_wide = TRUE, return_message = TRUE)
}
\arguments{
\item{return_wide}{\code{TRUE} return data in wide form}

\item{return_message}{\code{TRUE} return a message after data import}
}
\value{
\code{tibble}
}
\description{
This function returns the most recent
index values for the
41 primary MSCI indicies on a 15 minute time delay.
}
\examples{
msci_realtime_index_values()
}
\references{
\href{http://msci.com}{MSCI Inc}
}
\seealso{
Other MSCI: 
\code{\link{msci_indicies}()},
\code{\link{msci_indicies_constituents}()}

Other index values: 
\code{\link{crsp_indicies_returns}()},
\code{\link{fred_symbols}()},
\code{\link{nareit_annual_subsector_returns}()},
\code{\link{nareit_monthly_returns}()}

Other real-time data: 
\code{\link{dtcc_recent_trades}()},
\code{\link{us_public_companies}()}
}
\concept{MSCI}
\concept{index values}
\concept{real-time data}
