% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sec_2.R
\name{munge_tbl}
\alias{munge_tbl}
\title{Munge a tibble}
\usage{
munge_tbl(
  data,
  snake_names = F,
  unformat = F,
  convert_case = T,
  amount_digits = 2,
  include_address = T
)
}
\arguments{
\item{data}{a \code{tibble}}

\item{snake_names}{if \code{TRUE} returns snake case names}

\item{unformat}{if \code{TRUE} no formattable digits}

\item{convert_case}{if \code{TRUE} normalizes non url character columns to upper}

\item{amount_digits}{formattable digits}

\item{include_address}{if \code{TRUE} builds addresses}
}
\description{
Munge a tibble
}
