% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nareit.R
\name{nareit_capital_raises}
\alias{nareit_capital_raises}
\title{NAREIT capital raises}
\usage{
nareit_capital_raises(capital_type = NULL, nest_data = TRUE,
  return_message = TRUE)
}
\arguments{
\item{capital_type}{form of capital raised \itemize{
\item \code{NULL}: returns all types (default)
\item \code{IPO}: Initial Public Offering
\item \code{SECONDARY} :Secondary Equity Offering
\item \code{DEBT}: Debt Issuance
}}

\item{nest_data}{\code{TRUE} return nested data frame}

\item{return_message}{\code{TRUE} return a message after data import}
}
\value{
nested \code{tibble} or \code{tibble} if \code{nest_data = FALSE}#' @export
}
\description{
This function acquires data on REIT
capital raises by form of capital since 1989.
}
\examples{
\dontrun{
nareit_capital_raises(capital_type = NULL, nest_data = FALSE, return_message = TRUE)
}

}
\seealso{
Other NAREIT: \code{\link{nareit_annual_subsector_returns}},
  \code{\link{nareit_constituent_years}},
  \code{\link{nareit_entities}},
  \code{\link{nareit_industry_tracker}},
  \code{\link{nareit_mergers_acquisitions}},
  \code{\link{nareit_monthly_returns}},
  \code{\link{nareit_notable_properties}},
  \code{\link{nareit_property_msa}},
  \code{\link{nareit_state_info}}, \code{\link{reit_funds}}
}
\concept{NAREIT}
\concept{capital raises}
