% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nareit.R
\name{nareit_industry_tracker}
\alias{nareit_industry_tracker}
\title{NAREIT quarterly industry tracking data}
\usage{
nareit_industry_tracker(
  nest_data = FALSE,
  return_wide = TRUE,
  return_message = TRUE
)
}
\arguments{
\item{nest_data}{\code{TRUE} return nested data frame}

\item{return_wide}{\code{TRUE} return data in wide form}

\item{return_message}{\code{TRUE} return a message after data import}
}
\value{
nested \code{tibble} or \code{tibble} if \code{nest_data = FALSE}
}
\description{
This function acquires data on NAREIT
performance metrics including Funds From Operation [FFO],
Dividends, and Net Operating Income from 2000.
}
\examples{
\dontrun{
nareit_industry_tracker(return_message = TRUE, return_wide = TRUE)
}
}
\seealso{
Other NAREIT: 
\code{\link{nareit_annual_subsector_returns}()},
\code{\link{nareit_capital_raises}()},
\code{\link{nareit_constituent_years}()},
\code{\link{nareit_entities}()},
\code{\link{nareit_mergers_acquisitions}()},
\code{\link{nareit_monthly_returns}()},
\code{\link{nareit_notable_properties}()},
\code{\link{nareit_property_msa}()},
\code{\link{nareit_state_info}()},
\code{\link{reit_funds}()}
}
\concept{NAREIT}
