% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nareit.R
\name{nareit_mergers_acquisitions}
\alias{nareit_mergers_acquisitions}
\title{NAREIT merger and acquisitions}
\usage{
nareit_mergers_acquisitions(
  pages = 32:33,
  nest_data = FALSE,
  return_message = TRUE
)
}
\arguments{
\item{pages}{Integer vector. Page numbers to OCR from the PDF (default 32:33).}

\item{nest_data}{Logical. If \code{TRUE}, returns nested data frame.}

\item{return_message}{Logical. If \code{TRUE}, returns progress messages.}
}
\value{
A tibble containing NAREIT M&A activity data.
}
\description{
This function OCRs data on NAREIT member
M&A activity since 2004.
}
\examples{
\dontrun{
nareit_mergers_acquisitions(nest_data = FALSE, return_message = TRUE)
}
}
\seealso{
Other NAREIT: 
\code{\link{nareit_annual_subsector_returns}()},
\code{\link{nareit_capital_raises}()},
\code{\link{nareit_constituent_years}()},
\code{\link{nareit_entities}()},
\code{\link{nareit_industry_tracker}()},
\code{\link{nareit_monthly_returns}()},
\code{\link{nareit_notable_properties}()},
\code{\link{nareit_property_msa}()},
\code{\link{nareit_state_info}()},
\code{\link{reit_funds}()}

Other transaction data: 
\code{\link{dtcc_recent_trades}()},
\code{\link{dtcc_trades}()},
\code{\link{sec_securities_metrics_by_exchange}()}
}
\concept{NAREIT}
\concept{transaction data}
