% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nareit.R
\name{nareit_monthly_returns}
\alias{nareit_monthly_returns}
\title{NAREIT monthly returns by sector}
\usage{
nareit_monthly_returns(return_wide = TRUE, return_message = TRUE)
}
\arguments{
\item{return_wide}{\code{TRUE} return data in wide form}

\item{return_message}{\code{TRUE} return a message after data import}
}
\value{
\code{data_frame}
}
\description{
This function acquires FTSE NAREIT monthly return data
by sector from 1971 to the prior month's end.
}
\examples{
\dontrun{
nareit_monthly_returns(return_wide = FALSE)
}
}
\references{
\href{http://nareit.org}{National Association of Real Estate Investment Trusts}
}
\seealso{
Other NAREIT: \code{\link{nareit_annual_subsector_returns}},
  \code{\link{nareit_capital_raises}},
  \code{\link{nareit_constituent_years}},
  \code{\link{nareit_entities}},
  \code{\link{nareit_industry_tracker}},
  \code{\link{nareit_mergers_acquisitions}},
  \code{\link{nareit_notable_properties}},
  \code{\link{nareit_property_msa}},
  \code{\link{nareit_state_info}}, \code{\link{reit_funds}}

Other index values: \code{\link{crsp_indicies_returns}},
  \code{\link{fred_symbols}},
  \code{\link{msci_realtime_index_values}},
  \code{\link{nareit_annual_subsector_returns}}
}
\concept{NAREIT}
\concept{index values}
