% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nareit.R
\name{nareit_notable_properties}
\alias{nareit_notable_properties}
\title{NAREIT notable properties}
\usage{
nareit_notable_properties()
}
\value{
\code{tibble}
}
\description{
This function returns information about notable NAREIT identified properties.
}
\examples{
nareit_notable_properties()
}
\references{
\href{http://nareit.org}{National Association of Real Estate Investment Trusts}
}
\seealso{
Other NAREIT: \code{\link{nareit_annual_subsector_returns}},
  \code{\link{nareit_capital_raises}},
  \code{\link{nareit_constituent_years}},
  \code{\link{nareit_entities}},
  \code{\link{nareit_industry_tracker}},
  \code{\link{nareit_mergers_acquisitions}},
  \code{\link{nareit_monthly_returns}},
  \code{\link{nareit_property_msa}},
  \code{\link{nareit_state_info}}, \code{\link{reit_funds}}

Other property data: \code{\link{nareit_property_msa}},
  \code{\link{nareit_state_info}}
}
\concept{NAREIT}
\concept{property data}
