% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nareit.R
\name{nareit_property_msa}
\alias{nareit_property_msa}
\title{NAREIT properties by MSA}
\usage{
nareit_property_msa(nest_data = TRUE, return_message = TRUE)
}
\arguments{
\item{nest_data}{\code{TRUE} return nested data frame}

\item{return_message}{\code{TRUE} return a message after data import}
}
\value{
nested \code{tibble} or \code{tibble} if \code{nest_data = FALSE}
}
\description{
This function returns aggregated REIT owned
properties by Metropolitan Statistical Area [MSA].
}
\examples{
\dontrun{
nareit_property_msa(nest_data = TRUE, return_message = TRUE)
}
}
\references{
\href{http://nareit.org}{National Association of Real Estate Investment Trusts}
}
\seealso{
Other NAREIT: \code{\link{nareit_annual_subsector_returns}},
  \code{\link{nareit_capital_raises}},
  \code{\link{nareit_constituent_years}},
  \code{\link{nareit_entities}},
  \code{\link{nareit_industry_tracker}},
  \code{\link{nareit_mergers_acquisitions}},
  \code{\link{nareit_monthly_returns}},
  \code{\link{nareit_notable_properties}},
  \code{\link{nareit_state_info}}, \code{\link{reit_funds}}

Other property data: \code{\link{nareit_notable_properties}},
  \code{\link{nareit_state_info}}
}
\concept{NAREIT}
\concept{property data}
