% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nareit.R
\name{nareit_state_info}
\alias{nareit_state_info}
\title{NAREIT properties by state}
\usage{
nareit_state_info(include_reit_hq = TRUE, include_reit_holdings = TRUE,
  nest_data = TRUE, return_message = TRUE)
}
\arguments{
\item{include_reit_hq}{\code{TRUE} include REIT headquarter locations}

\item{include_reit_holdings}{\code{TRUE} include REIT aggregate holdings}

\item{nest_data}{\code{TRUE} return nested data frame}

\item{return_message}{\code{TRUE} return a message after data import}
}
\value{
nested \code{data_frame} or \code{data_frame} if \code{nest_data = FALSE}
}
\description{
This function returns information about REIT owned
holdings by state.
}
\examples{
\dontrun{
nareit_state_info(include_reit_hq = TRUE, include_reit_holdings = TRUE, nest_data = TRUE, return_message = TRUE)
}
}
\references{
\href{http://nareit.org}{National Association of Real Estate Investment Trusts}
}
\seealso{
Other NAREIT: \code{\link{nareit_annual_subsector_returns}},
  \code{\link{nareit_capital_raises}},
  \code{\link{nareit_constituent_years}},
  \code{\link{nareit_entities}},
  \code{\link{nareit_industry_tracker}},
  \code{\link{nareit_mergers_acquisitions}},
  \code{\link{nareit_monthly_returns}},
  \code{\link{nareit_notable_properties}},
  \code{\link{nareit_property_msa}},
  \code{\link{reit_funds}}

Other property data: \code{\link{nareit_notable_properties}},
  \code{\link{nareit_property_msa}}
}
\concept{NAREIT}
\concept{property data}
