% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf_functions.R
\name{parse_for_currency_value}
\alias{parse_for_currency_value}
\title{Parse string to currency value}
\usage{
parse_for_currency_value(x)
}
\arguments{
\item{x}{Character string containing a currency value (e.g., "$1,000,000")}
}
\value{
A \code{formattable::currency} object
}
\description{
Converts a character string containing currency notation to a
numeric currency value using formattable.
}
\seealso{
Other parsing: 
\code{\link{parse_for_percentage}()},
\code{\link{parse_multiple}()}
}
\concept{parsing}
\keyword{internal}
