% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf_functions.R
\name{parse_multiple}
\alias{parse_multiple}
\title{Parse string to numeric multiple}
\usage{
parse_multiple(x)
}
\arguments{
\item{x}{Character or numeric value representing a multiple}
}
\value{
A numeric value
}
\description{
Extracts a numeric value from a string representing a multiple
(e.g., "10x" becomes 10).
}
\seealso{
Other parsing: 
\code{\link{parse_for_currency_value}()},
\code{\link{parse_for_percentage}()}
}
\concept{parsing}
\keyword{internal}
