% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debt_functions.R
\name{parse_settlement_date}
\alias{parse_settlement_date}
\title{Parse Settlment date}
\usage{
parse_settlement_date(date = "2016-09-01",
  settlement_id = "2_PRIOR_LAST_WED")
}
\arguments{
\item{settlement_id}{}
}
\description{
Parse Settlment date
}

