% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaob.R
\name{pcaob_auditors}
\alias{pcaob_auditors}
\title{PCAOB Audit Relationships}
\usage{
pcaob_auditors(include_denied_firms = TRUE)
}
\arguments{
\item{include_denied_firms}{adds PCAOB denied firm feature}
}
\value{
a \code{tibble} with PCAOB auditor relationships and firm filings
}
\description{
PCAOB Audit Relationships
}
\examples{
\dontrun{
pcaob_auditors(include_denied_firms = TRUE)
}
}
