% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaob.R
\name{pcaob_denied_firms}
\alias{pcaob_denied_firms}
\title{Firms Denying PCAOB Audits}
\usage{
pcaob_denied_firms()
}
\value{
a \code{tibble} with firms that have denied PCAOB audit access
}
\description{
Firms Denying PCAOB Audits
}
\examples{
\dontrun{
pcaob_denied_firms()
}
}
