% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fred_api.R
\name{plot_time_series_static}
\alias{plot_time_series_static}
\title{Plot ggplot time series}
\usage{
plot_time_series_static(data, date_start = NULL, date_end = NULL,
  transformations = c("mean", "median", "smooth"),
  include_recessions = TRUE, x_breaks = 12, plot_labels = TRUE,
  use_hrbr_theme = TRUE)
}
\arguments{
\item{data}{a data frame}

\item{date_start}{if not NULL date starting}

\item{date_end}{if not NULL date ending}

\item{transformations}{\itemize{
\item \code{smooth} -- smoothed line
\item \code{mean} -- mean
\item \code{median}
}}

\item{include_recessions}{if \code{TRUE} includes recession bars}

\item{x_breaks}{number of x breks}

\item{plot_labels}{if \code{TRUE} plots labels}

\item{use_hrbr_theme}{if \code{TRUE} returns hrbr theme}
}
\value{
a ggplot2 object
}
\description{
Plot ggplot time series
}
