% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fred_api.R
\name{read_rda_file}
\alias{read_rda_file}
\title{read RDA file}
\usage{
read_rda_file(
  file =
    "https://github.com/abresler/FRED_Dictionaries/blob/master/data/fred_series_data.rda?raw=true",
  return_tibble = TRUE
)
}
\arguments{
\item{file}{path to RDA file or URL}

\item{return_tibble}{if \code{TRUE} returns a tibble}
}
\value{
a \code{tibble} or data frame with the loaded RDA data
}
\description{
read RDA file
}
\examples{
\dontrun{
read_rda_file(file = "path/to/file.rda")
}
}
