% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgar_rf.R
\name{recent_insider_trades}
\alias{recent_insider_trades}
\title{Insider trades, most recent}
\usage{
recent_insider_trades(nest_data = FALSE, return_message = TRUE)
}
\arguments{
\item{nest_data}{\code{TRUE} return nested data frame}

\item{return_message}{\code{TRUE} return a message after data import}
}
\value{
nested \code{tibble} or \code{tibble} if \code{nest_data = FALSE}
}
\description{
This function returns the most recent insider
trades as filed with the SEC.
}
\examples{
recent_insider_trades(nest_data = TRUE)
}
\seealso{
Other SEC: 
\code{\link{data_rf_leis}()},
\code{\link{dictionary_sec_filing_codes}()},
\code{\link{dictionary_sec_form_codes}()},
\code{\link{dictionary_sic_codes}()},
\code{\link{edgar_filing_streams}()},
\code{\link{mmf_owned_debt_securities}()},
\code{\link{rf_sec_13F_companies}()},
\code{\link{sec_bankruptcies}()},
\code{\link{sec_broker_dealers}()},
\code{\link{sec_ciks}()},
\code{\link{sec_closed_end_funds}()},
\code{\link{sec_cusips}()},
\code{\link{sec_failed_to_deliver_securities}()},
\code{\link{sec_filer}()},
\code{\link{sec_filing_entities}()},
\code{\link{sec_foia_requests}()},
\code{\link{sec_investment_companies}()},
\code{\link{sec_money_market_funds}()},
\code{\link{sec_municipal_advisors}()},
\code{\link{sec_securities_filing_counts}()},
\code{\link{sec_securities_metrics_by_exchange}()},
\code{\link{sec_xbrl_periods}()},
\code{\link{securities_offerings}()},
\code{\link{us_public_companies}()}

Other Rank and Filed: 
\code{\link{data_rf_leis}()},
\code{\link{mmf_owned_debt_securities}()},
\code{\link{rf_sec_13F_companies}()},
\code{\link{sec_filer}()},
\code{\link{sec_filing_entities}()},
\code{\link{securities_offerings}()},
\code{\link{us_public_companies}()}
}
\concept{Rank and Filed}
\concept{SEC}
\concept{securities transaction}
