% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_residual_calcs.R
\name{residual_values}
\alias{residual_values}
\title{Residual value calculations}
\usage{
residual_values(
  net_operating_income = 10,
  exit_cap_rate = 0.08,
  cost_of_sale = 0.05,
  debt = 0,
  sale_date = Sys.Date(),
  has_guarantee = FALSE,
  return_message = TRUE
)
}
\arguments{
\item{net_operating_income}{vector of net operating incomes}

\item{exit_cap_rate}{vector of exit cap rates}

\item{cost_of_sale}{vector of cost of sale}

\item{debt}{vector of debt amounts}

\item{sale_date}{date of the sale}

\item{has_guarantee}{if \code{TRUE} indicates a guarantee exists}

\item{return_message}{if \code{TRUE} returns a message}
}
\value{
a \code{tibble} with residual value calculations
}
\description{
Residual value calculations
}
\examples{
\dontrun{
residual_values(net_operating_income = 10, exit_cap_rate = 0.08)
}
}
