% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adv_functions.R
\name{sec_adv_manager_sitemap}
\alias{sec_adv_manager_sitemap}
\title{Returns data frame of SEC ADV sitemap}
\usage{
sec_adv_manager_sitemap()
}
\value{
A tibble containing SEC ADV section identifiers and their corresponding names
}
\description{
Returns data frame of SEC ADV sitemap
}
\examples{
\dontrun{
sec_adv_manager_sitemap()
}
}
