% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgar_rf.R
\name{sec_filer}
\alias{sec_filer}
\title{SEC filer}
\usage{
sec_filer(entity_names = NULL, tickers = NULL, ciks = NULL,
  tables = NULL, nest_data = FALSE,
  parse_all_filing_url_data = FALSE, parse_xbrl = FALSE,
  parse_subsidiaries = FALSE, parse_13F = FALSE,
  parse_asset_files = FALSE, parse_small_offerings = FALSE,
  parse_complete_text_filings = FALSE, parse_form_d = FALSE,
  parse_form_3_4s = FALSE, assign_to_environment = TRUE,
  return_message = TRUE)
}
\arguments{
\item{entity_names}{vector names to search}

\item{tickers}{character vector of ticker symbols to search}

\item{ciks}{numeric vector of CIKs}

\item{tables}{tables to include if they exist \itemize{
\code{NULL, General, CIK Filings, Filings, Private Offerings, Related Parties, Traders, C Level, MDA, Owners, Insider Trades, Trades}
\code{NULL}: selects all tables
\item \code{General}: general information about the filer
\item \code{CIK Filings}: summarised filings for a CIK
\item \code{Filings}: summarised filings for an entity, slightly different than \code{CIK Filings}
\item \code{Private Offerings}: parses any private offerings
\item \code{Related Parties}:  parses any related parties [people]
\item \code{Traders}: parses major traders
\item \code{C Level}: parses information about executives
\item \code{MDA}: parses text from company 10-K Management Discussion and Analysis [MDA] section
\item \code{Owners}: parses information about major owners
\item \code{Insider Trades}: parses insider trade information
\item \code{Trades}: parses all trade information
}}

\item{nest_data}{return a nested data frame \code{TRUE, FALSE}}

\item{parse_all_filing_url_data}{\code{TRUE} parses every SEC fling link}

\item{parse_xbrl}{\code{TRUE} parse XBRL for public companies, data starts in 2009}

\item{parse_subsidiaries}{\code{TRUE} parse all filer subsidiaries (default)}

\item{parse_13F}{\code{TRUE} parse \href{https://en.wikipedia.org/wiki/Form_13F}{13F's} for institutional managers}

\item{parse_asset_files}{\code{TRUE} parses ABS XML for \href{https://www.sec.gov/info/edgar/specifications/absxml.htm}{ABS Asset Data}
filing entities (default)}

\item{parse_small_offerings}{\code{TRUE} parses \href{https://www.sec.gov/info/smallbus/secg/rccomplianceguide-051316.htm}{Regulation CrowdFunding}
Form 1-A data if any exists for a filer (default)}

\item{assign_to_environment}{\code{true} assigns individual data frames to your environment}

\item{return_message}{\code{TRUE} return a message after data import}
}
\value{
where \code{nest_data} is \code{TRUE} a nested data_frame by asset,
where \code{nest_data} is \code{FALSE} a data_frame
}
\description{
This is function imports data
for a specified SEC filing entity.  An
SEC filing entity can be a person, public company or
private filer.  This function requires that the entity has a
Central Index Key [CIK].
}
\details{
The function acquires information for the
specified tables and will auto parse forms if the filer has them and
the user activates these parameters.
}
\examples{
\dontrun{
sec_filer(entity_names = 'HLT Holdco', tickers = c('FB'),
nest_data = TRUE, parse_subsidiaries = TRUE, parse_all_filing_url_data = TRUE,
parse_13F = TRUE, assign_to_environment = TRUE,
return_message = TRUE)

## Small Asset Filer Example

## ABS Example

#XBRL Example

}
}
\seealso{
Other SEC: \code{\link{data_rf_leis}},
  \code{\link{dictionary_sec_filing_codes}},
  \code{\link{dictionary_sec_form_codes}},
  \code{\link{dictionary_sic_codes}},
  \code{\link{edgar_filing_streams}},
  \code{\link{mmf_owned_debt_securities}},
  \code{\link{recent_insider_trades}},
  \code{\link{rf_sec_13F_companies}},
  \code{\link{sec_bankruptcies}},
  \code{\link{sec_broker_dealers}}, \code{\link{sec_ciks}},
  \code{\link{sec_closed_end_funds}},
  \code{\link{sec_cusips}},
  \code{\link{sec_failed_to_deliver_securities}},
  \code{\link{sec_filing_entities}},
  \code{\link{sec_foia_requests}},
  \code{\link{sec_investment_companies}},
  \code{\link{sec_money_market_funds}},
  \code{\link{sec_municipal_advisors}},
  \code{\link{sec_securities_filing_counts}},
  \code{\link{sec_securities_metrics_by_exchange}},
  \code{\link{securities_offerings}},
  \code{\link{us_public_companies}}

Other Rank and Filed: \code{\link{data_rf_leis}},
  \code{\link{mmf_owned_debt_securities}},
  \code{\link{recent_insider_trades}},
  \code{\link{rf_sec_13F_companies}},
  \code{\link{sec_filing_entities}},
  \code{\link{securities_offerings}},
  \code{\link{us_public_companies}}

Other entity search: \code{\link{adv_managers_current_period_summary}},
  \code{\link{adv_managers_filings}},
  \code{\link{adv_managers_metadata}},
  \code{\link{adv_managers_periods_summaries}},
  \code{\link{data_rf_leis}}, \code{\link{finra_entities}},
  \code{\link{nareit_entities}}, \code{\link{reit_funds}},
  \code{\link{rf_sec_13F_companies}},
  \code{\link{sec_bankruptcies}},
  \code{\link{sec_broker_dealers}}, \code{\link{sec_ciks}},
  \code{\link{sec_closed_end_funds}},
  \code{\link{sec_cusips}},
  \code{\link{sec_filing_entities}},
  \code{\link{sec_investment_companies}},
  \code{\link{sec_money_market_funds}},
  \code{\link{sec_municipal_advisors}},
  \code{\link{securities_offerings}},
  \code{\link{us_public_companies}},
  \code{\link{ycombinator_alumni}}

Other fund search: \code{\link{sec_investment_companies}}
}
\concept{Rank and Filed}
\concept{SEC}
\concept{XBRL}
\concept{entity search}
\concept{fund search}
