% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sec_2.R
\name{sec_filing_streams_rf}
\alias{sec_filing_streams_rf}
\title{SEC filing stream}
\usage{
sec_filing_streams_rf(
  filers = c("All", "Corporate Insider", "Companies", "Investment Company"),
  filing_names = c("All", "Annual Reports", "Quarterly Reports", "Current Reports",
    "Other Reports", "Registrations", "Private Offerings", "Ownership", "Prospectuses",
    "Exemptions", "Withdrawals", "Correspondence", "Proxy Statements", "Confidential"),
  nest_data = TRUE,
  return_message = TRUE
)
}
\arguments{
\item{filers}{type of filer \itemize{
\item \code{All}: all filer types (default)
\item \code{Corporate Insider}: corporate insiders
\item \code{Companies}: companies
\item \code{Investment Company}: acquires investment company data
}}

\item{filing_names}{type of filing \itemize{
\item \code{Annual Reports}: annual report
\item \code{Quarterly Reports}: quarterly report
\item \code{Current Reports}: current report
\item \code{Other Reports}: other reports
\item \code{Registrations}: securities registration
\item \code{Private Offerings}: private securities offerings
\item \code{Ownership}: ownership
\item \code{Prospectuses}: securities prospectus
\item \code{Exemptions}: exempt securities
\item \code{Withdrawals}: securities withdrawls
\item \code{Correspondence}: SEC correspondence
\item \code{Proxy Statements}:  proxy issuances
\item \code{Confidential}: confidential information
}}

\item{nest_data}{Logical. If \code{TRUE} (default), returns nested tibble.}

\item{return_message}{Logical. If \code{TRUE}, returns progress messages.}
}
\value{
A tibble containing SEC filing stream data grouped by filer and filing type.
}
\description{
This function returns the most recent SEC filings
for specified filer type by filing type.
}
\examples{
\dontrun{
sec_filing_streams_rf(filers = 'All', filing_names = 'Annual Reports')
}

}
