% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgar_rf.R
\name{sec_form_ds}
\alias{sec_form_ds}
\title{SEC Form-D's}
\usage{
sec_form_ds(
  industries = NULL,
  form_years = NULL,
  months = NULL,
  nest_data = FALSE,
  return_message = TRUE
)
}
\arguments{
\item{industries}{industries to search options: \itemize{
\item \code{NULL}: returns all industries(default)
\item \code{ENERGY}: energy sector
\item \code{FINANCIAL}: financial sector
\item \code{HEALTHCARE}: health-care sector
\item \code{OTHER}: all other sectors
\item \code{REAL ESTATE}: real estate sector
\item \code{TECHNOLOGY}: technology sector
\item \code{TRAVEL AND LEISURE}: travel and leisure sector
}}

\item{form_years}{years to search options are \itemize{
\item \code{NULL}: all years (default)
\item \code{2009-present}: search years
}}

\item{months}{months to search \itemize{
\item \code{NULL}: all months (default)
\code{1:12}: numeric month
}}

\item{nest_data}{\code{TRUE} return nested data frame}

\item{return_message}{\code{TRUE} return a message after data import}
}
\value{
where \code{nest_data} is \code{TRUE} a nested tibble by asset,
where \code{nest_data} is \code{FALSE} a tibble
}
\description{
This function returns data for SEC Form D's by
specified industry
}
\details{
\code{sec_form_ds()} queries all SEC filed form-d's since 2009 and returns the associated data.
the default parameters search every industry and year which you
can change by modifying the parameters
}
\examples{
\dontrun{
sec_form_ds()
sec_form_ds(form_years = 2016:2017, industries = c("Real Estate", "Technology", "Other"))
}
}
