% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sec_functions.R
\name{sec_money_market_funds}
\alias{sec_money_market_funds}
\title{SEC registered money market funds}
\usage{
sec_money_market_funds(
  only_most_recent = TRUE,
  years = NULL,
  months = NULL,
  nest_data = TRUE,
  return_message = TRUE
)
}
\arguments{
\item{only_most_recent}{\code{TRUE} search only the most recent period}

\item{years}{years to include starting in 2013}

\item{months}{months to include}

\item{nest_data}{\code{TRUE} return nested data frame}

\item{return_message}{\code{TRUE} return a message after data import}
}
\value{
nested \code{tibble} or \code{tibble} if \code{nest_data = FALSE}


}
\description{
This function returns information about SEC registered
money market funds for specified months and years from
January 2013 onwards.
}
\examples{
\dontrun{
sec_money_market_funds(only_most_recent = TRUE, nest_data = FALSE)
sec_money_market_funds(only_most_recent = FALSE, years = 2013:2016)
}
}
\references{
\href{http://sec.gov}{The Securities and Exchange Commission}
}
\seealso{
Other SEC: 
\code{\link{data_rf_leis}()},
\code{\link{dictionary_sec_filing_codes}()},
\code{\link{dictionary_sec_form_codes}()},
\code{\link{dictionary_sic_codes}()},
\code{\link{edgar_filing_streams}()},
\code{\link{mmf_owned_debt_securities}()},
\code{\link{recent_insider_trades}()},
\code{\link{rf_sec_13F_companies}()},
\code{\link{sec_bankruptcies}()},
\code{\link{sec_broker_dealers}()},
\code{\link{sec_ciks}()},
\code{\link{sec_closed_end_funds}()},
\code{\link{sec_cusips}()},
\code{\link{sec_failed_to_deliver_securities}()},
\code{\link{sec_filer}()},
\code{\link{sec_filing_entities}()},
\code{\link{sec_foia_requests}()},
\code{\link{sec_investment_companies}()},
\code{\link{sec_municipal_advisors}()},
\code{\link{sec_securities_filing_counts}()},
\code{\link{sec_securities_metrics_by_exchange}()},
\code{\link{sec_xbrl_periods}()},
\code{\link{securities_offerings}()},
\code{\link{us_public_companies}()}

Other entity search: 
\code{\link{adv_managers_current_period_summary}()},
\code{\link{adv_managers_filings}()},
\code{\link{adv_managers_metadata}()},
\code{\link{adv_managers_periods_summaries}()},
\code{\link{data_rf_leis}()},
\code{\link{finra_entities}()},
\code{\link{nareit_entities}()},
\code{\link{reit_funds}()},
\code{\link{rf_sec_13F_companies}()},
\code{\link{sec_bankruptcies}()},
\code{\link{sec_broker_dealers}()},
\code{\link{sec_ciks}()},
\code{\link{sec_closed_end_funds}()},
\code{\link{sec_cusips}()},
\code{\link{sec_filer}()},
\code{\link{sec_filing_entities}()},
\code{\link{sec_investment_companies}()},
\code{\link{sec_municipal_advisors}()},
\code{\link{sec_xbrl_periods}()},
\code{\link{securities_offerings}()},
\code{\link{us_public_companies}()},
\code{\link{ycombinator_alumni}()}

Other fund data: 
\code{\link{adv_managers_current_period_summary}()},
\code{\link{adv_managers_filings}()},
\code{\link{adv_managers_metadata}()},
\code{\link{adv_managers_periods_summaries}()},
\code{\link{reit_funds}()},
\code{\link{sec_ciks}()},
\code{\link{sec_closed_end_funds}()}
}
\concept{SEC}
\concept{entity search}
\concept{fund data}
