% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgar_rf.R
\name{sec_rules}
\alias{sec_rules}
\title{SEC Prefix Laws}
\usage{
sec_rules()
}
\value{
a  \code{tibble}
}
\description{
This function returns prefix
descriptions of SEC rules and the
law that created the rule
}
\seealso{
Other dictionary: 
\code{\link{dictionary_fred_ids}()},
\code{\link{dictionary_sec_filing_codes}()},
\code{\link{dictionary_sec_form_codes}()},
\code{\link{dictionary_sic_codes}()},
\code{\link{location_codes}()},
\code{\link{sic_naics_codes}()}
}
\concept{SEC
sec_rules()}
\concept{dictionary}
