% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgar_rf.R
\name{sec_securities_filing_counts}
\alias{sec_securities_filing_counts}
\title{SEC Filing Counts}
\usage{
sec_securities_filing_counts(return_message = TRUE)
}
\arguments{
\item{return_message}{\code{TRUE} return a message after data import}
}
\value{
a \code{tibble}
}
\description{
This function returns counts of SEC filings
by the type of form by year since 1994
}
\examples{
sec_securities_filing_counts()
}
\seealso{
Other SEC: \code{\link{data_rf_leis}},
  \code{\link{dictionary_sec_filing_codes}},
  \code{\link{dictionary_sec_form_codes}},
  \code{\link{dictionary_sic_codes}},
  \code{\link{edgar_filing_streams}},
  \code{\link{mmf_owned_debt_securities}},
  \code{\link{recent_insider_trades}},
  \code{\link{rf_sec_13F_companies}},
  \code{\link{sec_bankruptcies}},
  \code{\link{sec_broker_dealers}}, \code{\link{sec_ciks}},
  \code{\link{sec_closed_end_funds}},
  \code{\link{sec_cusips}},
  \code{\link{sec_failed_to_deliver_securities}},
  \code{\link{sec_filer}},
  \code{\link{sec_filing_entities}},
  \code{\link{sec_foia_requests}},
  \code{\link{sec_investment_companies}},
  \code{\link{sec_money_market_funds}},
  \code{\link{sec_municipal_advisors}},
  \code{\link{sec_securities_metrics_by_exchange}},
  \code{\link{securities_offerings}},
  \code{\link{us_public_companies}}
}
\concept{Rank and Fild}
\concept{SEC}
