% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sec_functions.R
\name{sec_xbrl_periods}
\alias{sec_xbrl_periods}
\title{SEC quarterly XBRL data}
\usage{
sec_xbrl_periods(
  only_most_recent = TRUE,
  years = NULL,
  quarters = NULL,
  tables = NULL,
  only_all = TRUE,
  assign_to_environment = TRUE,
  return_message = TRUE
)
}
\arguments{
\item{only_most_recent}{\code{TRUE} search only the most recent period}

\item{years}{years to include starting in 2009}

\item{quarters}{quarters to search \itemize{
\item \code{NULL}(default): all quarters,
\item \code{1}: Q1
\item \code{2}: Q2
\item \code{3}: Q3
\item \code{4}: Q4
}}

\item{tables}{underlying XBRL tables you want to include \itemize{
\item \code{NULL}: all tables (default)
\item \code{Presentation}: presentation data
\item \code{Values}: XBRL value data
\item \code{Filers}: filer data
\item \code{Taxonomy}: taxonomy data
}}

\item{only_all}{\code{TRUE} returns a joined file of the four tables}

\item{assign_to_environment}{\code{TRUE} assign the results into your environment}

\item{return_message}{\code{TRUE} return a message after data import}
}
\value{
nested \code{tibble} or \code{tibble}
}
\description{
This function returns quarterly full \href{https://en.wikipedia.org/wiki/XBRL}{XBRL} dumps
for all SEC registered public companies begining in 2009.
}
\examples{
\dontrun{
sec_xbrl_periods(only_most_recent = TRUE, only_all = TRUE, assign_to_environment = TRUE, return_message = TRUE)
}
}
\references{
\href{http://sec.gov}{The Securities and Exchange Commission}
}
\seealso{
Other SEC: 
\code{\link{data_rf_leis}()},
\code{\link{dictionary_sec_filing_codes}()},
\code{\link{dictionary_sec_form_codes}()},
\code{\link{dictionary_sic_codes}()},
\code{\link{edgar_filing_streams}()},
\code{\link{mmf_owned_debt_securities}()},
\code{\link{recent_insider_trades}()},
\code{\link{rf_sec_13F_companies}()},
\code{\link{sec_bankruptcies}()},
\code{\link{sec_broker_dealers}()},
\code{\link{sec_ciks}()},
\code{\link{sec_closed_end_funds}()},
\code{\link{sec_cusips}()},
\code{\link{sec_failed_to_deliver_securities}()},
\code{\link{sec_filer}()},
\code{\link{sec_filing_entities}()},
\code{\link{sec_foia_requests}()},
\code{\link{sec_investment_companies}()},
\code{\link{sec_money_market_funds}()},
\code{\link{sec_municipal_advisors}()},
\code{\link{sec_securities_filing_counts}()},
\code{\link{sec_securities_metrics_by_exchange}()},
\code{\link{securities_offerings}()},
\code{\link{us_public_companies}()}

Other public company data: 
\code{\link{sec_failed_to_deliver_securities}()},
\code{\link{sec_securities_metrics_by_exchange}()}

Other entity search: 
\code{\link{adv_managers_current_period_summary}()},
\code{\link{adv_managers_filings}()},
\code{\link{adv_managers_metadata}()},
\code{\link{adv_managers_periods_summaries}()},
\code{\link{data_rf_leis}()},
\code{\link{finra_entities}()},
\code{\link{nareit_entities}()},
\code{\link{reit_funds}()},
\code{\link{rf_sec_13F_companies}()},
\code{\link{sec_bankruptcies}()},
\code{\link{sec_broker_dealers}()},
\code{\link{sec_ciks}()},
\code{\link{sec_closed_end_funds}()},
\code{\link{sec_cusips}()},
\code{\link{sec_filer}()},
\code{\link{sec_filing_entities}()},
\code{\link{sec_investment_companies}()},
\code{\link{sec_money_market_funds}()},
\code{\link{sec_municipal_advisors}()},
\code{\link{securities_offerings}()},
\code{\link{us_public_companies}()},
\code{\link{ycombinator_alumni}()}
}
\concept{SEC}
\concept{entity search}
\concept{public company data}
