% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgar_rf.R
\name{securities_offerings}
\alias{securities_offerings}
\title{Securities offerings}
\usage{
securities_offerings(
  year_forms = NULL,
  forms = NULL,
  nest_data = FALSE,
  return_message = TRUE
)
}
\arguments{
\item{year_forms}{years to search starting in 1994, \code{NULL} returns all years}

\item{forms}{forms to use in search \itemize{
\item \code{NULL}: returns all (default)
\item \code{D}: Exempt offerings
\item \code{W}: Secondary sales
\item \code{S-1}: Initial public offerings
\item \code{S-4}: Initial public offering withdrawls
}}

\item{nest_data}{\code{TRUE} return nested data frame}

\item{return_message}{\code{TRUE} return a message after data import}
}
\value{
where \code{nest_data} is \code{TRUE} a nested tibble by asset,
where \code{nest_data} is \code{FALSE} a tibble
}
\description{
This function returns securities offering
data by form type and year from 1994 onward.
}
\examples{
#' \dontrun{

## All Securities Filings
securities_offerings(year_forms = NULL, forms = NULL, return_message = TRUE, nest_data = TRUE)

## IPOs since 1999
securities_offerings(year_forms = 1999:2017, forms = "S-1", return_message = TRUE, nest_data = FALSE)
}
}
\seealso{
Other SEC: 
\code{\link{data_rf_leis}()},
\code{\link{dictionary_sec_filing_codes}()},
\code{\link{dictionary_sec_form_codes}()},
\code{\link{dictionary_sic_codes}()},
\code{\link{edgar_filing_streams}()},
\code{\link{mmf_owned_debt_securities}()},
\code{\link{recent_insider_trades}()},
\code{\link{rf_sec_13F_companies}()},
\code{\link{sec_bankruptcies}()},
\code{\link{sec_broker_dealers}()},
\code{\link{sec_ciks}()},
\code{\link{sec_closed_end_funds}()},
\code{\link{sec_cusips}()},
\code{\link{sec_failed_to_deliver_securities}()},
\code{\link{sec_filer}()},
\code{\link{sec_filing_entities}()},
\code{\link{sec_foia_requests}()},
\code{\link{sec_investment_companies}()},
\code{\link{sec_money_market_funds}()},
\code{\link{sec_municipal_advisors}()},
\code{\link{sec_securities_filing_counts}()},
\code{\link{sec_securities_metrics_by_exchange}()},
\code{\link{sec_xbrl_periods}()},
\code{\link{us_public_companies}()}

Other entity search: 
\code{\link{adv_managers_current_period_summary}()},
\code{\link{adv_managers_filings}()},
\code{\link{adv_managers_metadata}()},
\code{\link{adv_managers_periods_summaries}()},
\code{\link{data_rf_leis}()},
\code{\link{finra_entities}()},
\code{\link{nareit_entities}()},
\code{\link{reit_funds}()},
\code{\link{rf_sec_13F_companies}()},
\code{\link{sec_bankruptcies}()},
\code{\link{sec_broker_dealers}()},
\code{\link{sec_ciks}()},
\code{\link{sec_closed_end_funds}()},
\code{\link{sec_cusips}()},
\code{\link{sec_filer}()},
\code{\link{sec_filing_entities}()},
\code{\link{sec_investment_companies}()},
\code{\link{sec_money_market_funds}()},
\code{\link{sec_municipal_advisors}()},
\code{\link{sec_xbrl_periods}()},
\code{\link{us_public_companies}()},
\code{\link{ycombinator_alumni}()}

Other securities search: 
\code{\link{mmf_owned_debt_securities}()},
\code{\link{rf_sec_13F_companies}()},
\code{\link{sec_cusips}()}

Other Rank and Filed: 
\code{\link{data_rf_leis}()},
\code{\link{mmf_owned_debt_securities}()},
\code{\link{recent_insider_trades}()},
\code{\link{rf_sec_13F_companies}()},
\code{\link{sec_filer}()},
\code{\link{sec_filing_entities}()},
\code{\link{us_public_companies}()}
}
\concept{Rank and Filed}
\concept{SEC}
\concept{entity search}
\concept{securities search}
