% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgar_rf.R
\name{sic_naics_codes}
\alias{sic_naics_codes}
\title{SIC Codes}
\usage{
sic_naics_codes(filter_duplicates = TRUE)
}
\arguments{
\item{filter_duplicates}{\code{TRUE} removes duplicate entries}
}
\value{
a \code{tibble}
}
\description{
This function returns Standard Industrial Classification [SIC]
codes and the corresponding
North American Industry Classification System [NAICS] identification.
}
\examples{
sic_naics_codes(filter_duplicates = TRUE)
}
\seealso{
Other dictionary: 
\code{\link{dictionary_fred_ids}()},
\code{\link{dictionary_sec_filing_codes}()},
\code{\link{dictionary_sec_form_codes}()},
\code{\link{dictionary_sic_codes}()},
\code{\link{location_codes}()},
\code{\link{sec_rules}()}
}
\concept{dictionary}
