% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicies.R
\name{sp500_constituents}
\alias{sp500_constituents}
\title{S&P 500 constituents}
\usage{
sp500_constituents(return_message = TRUE)
}
\arguments{
\item{return_message}{\code{TRUE} return a message after data import}
}
\value{
a \code{tibble}
}
\description{
This function returns the constituents
in the S&P 500.
}
\examples{
sp500_constituents(return_message = TRUE)
}
\seealso{
Other index constituents: \code{\link{crsp_indicies_constituents}},
  \code{\link{msci_indicies_constituents}},
  \code{\link{nareit_constituent_years}}
}
\concept{index constituents}
