% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trading_view.R
\name{tickers_news}
\alias{tickers_news}
\title{Tradingview tickers news}
\usage{
tickers_news(tickers = c("FB", "AAPL", "NFLX", "GOOG", "VNO", "EQR",
  "BXP"), return_message = TRUE, nest_data = FALSE)
}
\description{
Returns news data for specified tickers
}
\examples{
tickers_news(tickers = c("VNO", "AVB", "PEI"), return_message = TRUE, nest_data = FALSE)
}
