% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markit.R
\name{tickers_trades}
\alias{tickers_trades}
\title{Tickers trades}
\usage{
tickers_trades(tickers = c("PEI", "bxp"), return_message = TRUE)
}
\arguments{
\item{return_message}{}
}
\value{

}
\description{
Tickers trades
}
\examples{
tickers_trades(tickers = c("VNO", "BXP", "FB"))

}
