% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tidy_column_relations}
\alias{tidy_column_relations}
\title{Tidy columns with nested list or related columns}
\usage{
tidy_column_relations(data, column_keys = c("idCIK", "nameEntity"),
  bind_to_original_df = FALSE, clean_column_formats = TRUE)
}
\arguments{
\item{data}{data frame column}

\item{column_keys}{column keys}

\item{bind_to_original_df}{bind to the original data frame}

\item{clean_column_formats}{clean the columns}
}
\description{
Tidy columns with nested list or related columns
}

