% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterfall_functions.R
\name{tidy_promote_structure}
\alias{tidy_promote_structure}
\title{Tidy waterfall structure}
\usage{
tidy_promote_structure(
  promote_structures = c("20 over a 12", "30 / 18", "40 over a 10x"),
  return_wide = F
)
}
\arguments{
\item{promote_structures}{character vector of promote structures}

\item{return_wide}{return data in wide form}
}
\value{
a data frame
}
\description{
This function parses a character vector
describing a partnership waterfall into a
date frame.  The function looks to recogonize
whether the promote is based upon an internal rate of return
or capital multiple hurdle.
}
\examples{
tidy_promote_structure(promote_structures = c("20 over a 12", '30 / 18', "40 over a 10x"), return_wide = T)
}
\seealso{
Other calculation: 
\code{\link{calculate_cash_flow_dates}()},
\code{\link{calculate_cash_flow_waterfall_partnership}()},
\code{\link{calculate_cash_flow_waterfall}()},
\code{\link{calculate_cash_flows_returns}()},
\code{\link{calculate_days_accrued_pref}()},
\code{\link{calculate_irr_periods}()},
\code{\link{calculate_leverage_metrics}()},
\code{\link{calculate_loan_payment}()},
\code{\link{calculate_residual_valuation_cap_rates}()},
\code{\link{calculate_residual_valuation_ebitda_multiples}()},
\code{\link{calculate_share_proceeds}()},
\code{\link{calculate_valuation_post_money}()}

Other partnership calculation: 
\code{\link{calculate_cash_flow_dates}()},
\code{\link{calculate_cash_flow_waterfall_partnership}()},
\code{\link{calculate_cash_flow_waterfall}()},
\code{\link{calculate_cash_flows_returns}()},
\code{\link{calculate_days_accrued_pref}()},
\code{\link{calculate_irr_periods}()}
}
\concept{calculation}
\concept{partnership calculation}
