% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trading_view.R
\name{tv_market_events}
\alias{tv_market_events}
\title{Get Market Events}
\usage{
tv_market_events(return_message = TRUE)
}
\arguments{
\item{return_message}{if \code{TRUE} returns a message}
}
\value{
a \code{tibble} of market events
}
\description{
Returns a list of global financial
and economics events
}
\examples{
\dontrun{
tv_market_events(return_message = TRUE)
}
}
