% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trading_view.R
\name{tv_metric}
\alias{tv_metric}
\title{Generate tradeview metric query}
\usage{
tv_metric(
  filter = tibble(left = "market_cap_basic", operation = "nempty"),
  symbols = list(query = list(types = c("stock", "fund", "dr"))),
  metrics = c("change", "change_abs", "close", "description",
    "earnings_per_share_basic_ttm", "market_cap_basic", "name", "number_of_employees",
    "price_earnings_ttm", "sector", "volume"),
  sort = list(sortBy = "market_cap_basic", sortOrder = "desc"),
  options = list(lang = "eng"),
  range = c(0, 1500000)
)
}
\arguments{
\item{range}{}
}
\value{

}
\description{
Generate tradeview metric query
}
