% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trading_view.R
\name{tv_tickers_news}
\alias{tv_tickers_news}
\title{Tradingview tickers news}
\usage{
tv_tickers_news(
  tickers = c("FB", "AAPL", "NFLX", "GOOG", "VNO", "EQR", "BXP"),
  default_exchange = "NASDAQ",
  return_message = TRUE,
  nest_data = FALSE
)
}
\arguments{
\item{tickers}{vector of ticker symbols}

\item{return_message}{if \code{TRUE} returns a message}

\item{nest_data}{if \code{TRUE} nest data by ticker}
}
\value{
a \code{tibble} of news data
}
\description{
Returns news data for specified tickers
}
\examples{
\dontrun{
tv_tickers_news(tickers = c("VNO", "AVB", "PEI"), return_message = TRUE, nest_data = FALSE)
}
}
