% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/government_data.R
\name{us_banks}
\alias{us_banks}
\title{United States Bank Data}
\usage{
us_banks()
}
\value{
a \code{tibble} with US bank information from the FDIC
}
\description{
Information about United States Banks as monitored by the FDIC
}
\examples{
us_banks()
}
