% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inflation.R
\name{us_inflation}
\alias{us_inflation}
\title{US Inflation Adjustment Table}
\usage{
us_inflation(base_year = 1913)
}
\arguments{
\item{base_year}{1913 or 1800}
}
\value{
a \code{tibble}
}
\description{
Gets a table to adjust prices to current dollars
}
\examples{
us_inflation(base_year = 1913)
}
