% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treasury_auctions.R
\name{us_treasury_auctions}
\alias{us_treasury_auctions}
\title{Treasury Auction History}
\usage{
us_treasury_auctions()
}
\value{
a \code{tibble} of historical US treasury auction data
}
\description{
Treasury Auction History
}
\examples{
\dontrun{
us_treasury_auctions()
}
}
