% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_functions.R
\name{visualize_data_frame}
\alias{visualize_data_frame}
\title{Visualize a data frame}
\usage{
visualize_data_frame(data, edit = F)
}
\arguments{
\item{data}{A data frame}

\item{edit}{Do you wish to edit the data}
}
\description{
Visualize a data frame
}
\examples{
visualize_data_frame(mtcars)
}

