% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_functions.R
\name{visualize_tibble}
\alias{visualize_tibble}
\title{Data frame visualization}
\usage{
visualize_tibble(data, edit = F)
}
\arguments{
\item{data}{A data frame}

\item{edit}{\code{TRUE} edit the data frame}
}
\value{
\code{htmlwidget}
}
\description{
This function allows a data frame to
be viewed and edited via the \code{listviewer} package.
}
\examples{
visualize_tibble(mtcars)
}
