% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venture_data_functions.R
\name{ycombinator_alumni}
\alias{ycombinator_alumni}
\title{Y Combinator data}
\usage{
ycombinator_alumni(nest_data = FALSE, return_message = TRUE)
}
\arguments{
\item{nest_data}{\code{TRUE} return nested data frame}

\item{return_message}{\code{TRUE} return a message after data import
includes information about a random YC company}
}
\value{
where \code{nest_data} is \code{TRUE} a nested tibble by batch,
where \code{nest_data} is \code{FALSE} a tibble
}
\description{
This function returns information on
Y Combinator alumni and active particpants.
}
\examples{
ycombinator_alumni(nest_data = FALSE)
}
\references{
\href{https://www.ycombinator.com/}{YCombinator}
}
\seealso{
Other venture capital: \code{\link{calculate_valuation_post_money}},
  \code{\link{cb_unicorns}}

Other entity search: \code{\link{adv_managers_current_period_summary}},
  \code{\link{adv_managers_filings}},
  \code{\link{adv_managers_metadata}},
  \code{\link{adv_managers_periods_summaries}},
  \code{\link{data_rf_leis}}, \code{\link{finra_entities}},
  \code{\link{nareit_entities}}, \code{\link{reit_funds}},
  \code{\link{rf_sec_13F_companies}},
  \code{\link{sec_bankruptcies}},
  \code{\link{sec_broker_dealers}}, \code{\link{sec_ciks}},
  \code{\link{sec_closed_end_funds}},
  \code{\link{sec_cusips}}, \code{\link{sec_filer}},
  \code{\link{sec_filing_entities}},
  \code{\link{sec_investment_companies}},
  \code{\link{sec_money_market_funds}},
  \code{\link{sec_municipal_advisors}},
  \code{\link{sec_xbrl_periods}},
  \code{\link{securities_offerings}},
  \code{\link{us_public_companies}}
}
\concept{entity search}
\concept{venture capital}
