% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adv_functions.R
\name{adv_managers_periods_summaries}
\alias{adv_managers_periods_summaries}
\title{ADV managers periods data}
\usage{
adv_managers_periods_summaries(periods = c("2018-06"),
  all_periods = FALSE, only_most_recent = FALSE,
  include_exempt = TRUE, nest_data = FALSE, return_message = TRUE)
}
\arguments{
\item{periods}{dates in year-month form}

\item{all_periods}{include all periods}

\item{nest_data}{return a nested data frame}

\item{return_message}{return a message after parsing data}

\item{is_exempt}{exempt, non-exempt filers}
}
\value{
where \code{nest_data} is \code{TRUE} a nested tibble by period and type of filer,
where \code{nest_data} is \code{FALSE} a tibble
}
\description{
This function returns monthly summary
information for every ADV filing manager
from 2006 onwards.
}
\examples{
\dontrun{
adv_managers_periods_summaries(periods = c("2006-06", "2016-12", "2017-01"), all_periods = FALSE, is_exempt = c(FALSE,TRUE), only_most_recent = FALSE, nest_data = FALSE)

adv_managers_periods_summaries(only_most_recent = TRUE)
}
}
\seealso{
Other IAPD: \code{\link{adv_managers_current_period_summary}},
  \code{\link{adv_managers_filings}},
  \code{\link{adv_managers_metadata}},
  \code{\link{finra_entities}}, \code{\link{finra_people}}

Other ADV: \code{\link{adv_managers_current_period_summary}},
  \code{\link{adv_managers_filings}}

Other entity search: \code{\link{adv_managers_current_period_summary}},
  \code{\link{adv_managers_filings}},
  \code{\link{adv_managers_metadata}},
  \code{\link{data_rf_leis}}, \code{\link{finra_entities}},
  \code{\link{nareit_entities}}, \code{\link{reit_funds}},
  \code{\link{rf_sec_13F_companies}},
  \code{\link{sec_bankruptcies}},
  \code{\link{sec_broker_dealers}}, \code{\link{sec_ciks}},
  \code{\link{sec_closed_end_funds}},
  \code{\link{sec_cusips}}, \code{\link{sec_filer}},
  \code{\link{sec_filing_entities}},
  \code{\link{sec_investment_companies}},
  \code{\link{sec_money_market_funds}},
  \code{\link{sec_municipal_advisors}},
  \code{\link{sec_xbrl_periods}},
  \code{\link{securities_offerings}},
  \code{\link{us_public_companies}},
  \code{\link{ycombinator_alumni}}

Other fund data: \code{\link{adv_managers_current_period_summary}},
  \code{\link{adv_managers_filings}},
  \code{\link{adv_managers_metadata}},
  \code{\link{reit_funds}}, \code{\link{sec_ciks}},
  \code{\link{sec_closed_end_funds}},
  \code{\link{sec_money_market_funds}}
}
\concept{ADV}
\concept{IAPD}
\concept{entity search}
\concept{fund data}
