% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterfall_functions.R
\name{calculate_cash_flow_dates}
\alias{calculate_cash_flow_dates}
\title{Summary cash flows}
\usage{
calculate_cash_flow_dates(dates = c("2016-09-01", "2017-08-31", "2018-08-31",
  "2019-08-31", "2020-08-31", "2021-08-31", "2022-08-31", "2023-08-31"),
  cash_flows = c(-4151601, 119499.036215643, 257186.036215643,
  447646.036215643, 200652.036215643, 510409.036215643, 193.036215643166,
  8788626.76409155), working_capital = 125000,
  remove_cumulative_cols = TRUE, include_final_day = TRUE,
  distribution_frequency = NA)
}
\arguments{
\item{dates}{vector of dates}

\item{cash_flows}{vector of cash flows}

\item{working_capital}{amount of working capital, minimum cash}

\item{remove_cumulative_cols}{\code{TRUE} remove summary columns}

\item{include_final_day}{\code{TRUE} include the final day in calculation}

\item{distribution_frequency}{frequency of distribution \itemize{
\\item \code{NA}: NA
\item \code{weekly}: weekly distributions
\item \code{monthly}: monthly distributions
\item \code{quarterly}: quarterly distributions
\item \code{annually}: annual distributions
\item \code{sale}: distribution on residual
}}
}
\value{
data_frame
}
\description{
This function returns a data frame of
summarised cash flows for given a set of inputs.
}
\examples{
calculate_cash_flow_dates(dates = c( "2016-09-01", "2017-08-31", "2018-08-31", "2019-08-31", "2020-08-31", "2021-08-31", "2022-08-31", "2023-08-31" ),
cash_flows = c( -4151601, 119499.036215643, 257186.036215643, 447646.036215643, 200652.036215643, 510409.036215643, 193.036215643166, 8788626.7640915 ),
working_capital = 125000, remove_cumulative_cols = TRUE, include_final_day = TRUE, distribution_frequency = NA)
}
\seealso{
Other calculation: \code{\link{calculate_cash_flow_waterfall_partnership}},
  \code{\link{calculate_cash_flow_waterfall}},
  \code{\link{calculate_cash_flows_returns}},
  \code{\link{calculate_days_accrued_pref}},
  \code{\link{calculate_irr_periods}},
  \code{\link{calculate_leverage_metrics}},
  \code{\link{calculate_loan_payment}},
  \code{\link{calculate_residual_valuation_cap_rates}},
  \code{\link{calculate_residual_valuation_ebitda_multiples}},
  \code{\link{calculate_share_proceeds}},
  \code{\link{calculate_valuation_post_money}},
  \code{\link{tidy_promote_structure}}

Other leveraged finance calculation: \code{\link{calculate_cash_flow_waterfall_partnership}},
  \code{\link{calculate_cash_flow_waterfall}},
  \code{\link{calculate_cash_flows_returns}},
  \code{\link{calculate_irr_periods}},
  \code{\link{calculate_leverage_metrics}},
  \code{\link{calculate_loan_payment}}

Other partnership calculation: \code{\link{calculate_cash_flow_waterfall_partnership}},
  \code{\link{calculate_cash_flow_waterfall}},
  \code{\link{calculate_cash_flows_returns}},
  \code{\link{calculate_days_accrued_pref}},
  \code{\link{calculate_irr_periods}},
  \code{\link{tidy_promote_structure}}
}
\concept{calculation}
\concept{leveraged finance calculation}
\concept{partnership calculation}
