% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf_functions.R
\name{calculate_leverage_metrics}
\alias{calculate_leverage_metrics}
\title{Calculate leveraged return metrics}
\usage{
calculate_leverage_metrics(purchase_price = 0,
  capitalized_acquisition_costs = 0, capital_investment = 0, revenue = 0,
  expenses = 0, loan_to_cost = 0, interest_rate = 0,
  borrow_capital_investment = F, include_capitalized_cost = T,
  leverage_threshold = 0.95, is_interest_only = FALSE,
  interest_only_periods = 0, is_actual_360 = TRUE,
  amortization_years = 30, amortization_months = 0, term_years = 30,
  term_months = 0, pct_loan_fee = 0, return_wide = T,
  return_message = T)
}
\arguments{
\item{purchase_price}{Vector of Purchase Prices}

\item{capitalized_acquisition_costs}{Vector of Capitalized Acquisition Costs}

\item{capital_investment}{Vector of Capital Investment}

\item{revenue}{Vector of revenue amounts}

\item{expenses}{Vector of expenses}

\item{loan_to_cost}{Vector of loan to cost}

\item{interest_rate}{Interest Rate}

\item{borrow_capital_investment}{Borrow Investment \code{TRUE, FALSE}}

\item{include_capitalized_cost}{Include capitalized costs in leverage calculations \code{TRUE, FALSE}}

\item{leverage_threshold}{Maximum Leverage}

\item{is_interest_only}{Does loan have interst only periods \code{TRUE, FALSE}}

\item{interest_only_periods}{Interest Only Periods}

\item{is_actual_360}{Is loan calcuated on actual/360 basis \code{TRUE, FALSE}}

\item{amortization_years}{Loan amortization years}

\item{amortization_months}{Loan amortization months}

\item{term_years}{Term of the loan, years}

\item{term_months}{Term of the loan, months}

\item{pct_loan_fee}{Loan pee, percent}

\item{return_message}{}
}
\description{
Calculate leveraged return metrics
}

