% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf_functions.R
\name{calculate_share_proceeds}
\alias{calculate_share_proceeds}
\title{Calculate proceeds from share sale}
\usage{
calculate_share_proceeds(price, shares)
}
\arguments{
\item{shares}{}
}
\description{
Calculate proceeds from share sale
}

