% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sec_2.R
\name{cik_filing_counts}
\alias{cik_filing_counts}
\title{CIK Filing Counts}
\usage{
cik_filing_counts(cik, return_message = T)
}
\arguments{
\item{cik}{CIK codes}

\item{return_message}{}
}
\value{

}
\description{
CIK Filing Counts
}
