% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crisp.R
\name{crsp_indicies_constituents}
\alias{crsp_indicies_constituents}
\title{CRSP index constituents}
\usage{
crsp_indicies_constituents(nest_data = TRUE, return_message = TRUE)
}
\arguments{
\item{nest_data}{\code{TRUE} return nested data frame}

\item{return_message}{\code{TRUE} return a message after data import}
}
\value{
nested \code{tibble} or \code{tibble} if \code{nest_data = FALSE}
}
\description{
This function returns information
about a CRSP index's constituents
}
\examples{
crsp_indicies_constituents(nest_data = FALSE, return_message = TRUE)
}
\references{
\href{http://crsp.com}{The Center for Research in Security Prices}
}
\seealso{
Other CRSP: 
\code{\link{crsp_files}()},
\code{\link{crsp_indicies_returns}()}

Other index constituents: 
\code{\link{msci_indicies_constituents}()},
\code{\link{nareit_constituent_years}()},
\code{\link{sp500_constituents}()}
}
\concept{CRSP}
\concept{index constituents}
