% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dallas_fed.R
\name{dallas_fed_international_housing}
\alias{dallas_fed_international_housing}
\title{Dallas Federal Reserve International House Price Database}
\usage{
dallas_fed_international_housing(indicies = NULL, nest_data = FALSE,
  return_message = TRUE)
}
\arguments{
\item{return_message}{}
}
\description{
This retrieves data for the international house
price database maintained by the Dallas FED
}
\examples{
dallas_fed_international_housing()
}
