% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finbox.R
\name{dictionary_finbox_metrics}
\alias{dictionary_finbox_metrics}
\title{Get Searchable FinBOX metrics}
\usage{
dictionary_finbox_metrics()
}
\description{
Get Searchable FinBOX metrics
}
\examples{
dictionary_finbox_metrics()
}
