% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sec_2.R
\name{edgar_entities_cik}
\alias{edgar_entities_cik}
\title{SEC registered entity search}
\usage{
edgar_entities_cik(search_names, nest_data = FALSE, return_message = TRUE)
}
\arguments{
\item{return_message}{}
}
\description{
SEC registered entity search
}
\examples{
edgar_entities_cik( c("Rockwood", "BREA", 'EJF'))
}
