% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgar_rf.R
\name{edgar_entities_cik}
\alias{edgar_entities_cik}
\title{SEC registered entity search}
\usage{
edgar_entities_cik(search_names = c("Rockwood", "BREA", "EJF"),
  nest_data = FALSE, return_message = TRUE)
}
\arguments{
\item{return_message}{}
}
\description{
SEC registered entity search
}
