% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgar_rf.R
\name{edgar_ft_terms}
\alias{edgar_ft_terms}
\title{Title}
\usage{
edgar_ft_terms(search_terms = c("\\"Jared Kushner\\"",
  "\\"EJF Capital\\"", "\\"Blackstone Real Estate\\""),
  include_counts = TRUE, nest_data = FALSE, return_message = TRUE)
}
\description{
Title
}
\examples{
edgar_ft_terms(search_terms = c('"Jared Kushner"', '"EJF Capital"', '"Blackstone Real Estate"'))
}
