% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sec_2.R
\name{edgar_ft_terms}
\alias{edgar_ft_terms}
\title{Title}
\usage{
edgar_ft_terms(
  search_terms = c("\\"Jared Kushner\\"", "\\"EJF Capital\\"",
    "\\"Blackstone Real Estate\\""),
  include_counts = F,
  nest_data = FALSE,
  return_message = TRUE
)
}
\value{

}
\description{
Title
}
\examples{
edgar_ft_terms(search_terms = c('"Jared Kushner"', '"EJF Capital"', '"Blackstone Real Estate"'))
}
