% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgar_rf.R
\name{edgar_sic_filers}
\alias{edgar_sic_filers}
\title{SIC Cod Companies}
\usage{
edgar_sic_filers(sic_codes = c(3949, 3690, 3711), merge_names = TRUE,
  return_message = TRUE, nest_data = FALSE)
}
\arguments{
\item{nest_data}{}
}
\description{
SIC Cod Companies
}
