% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sec_2.R
\name{edgar_tickers}
\alias{edgar_tickers}
\title{SEC listed public companys}
\usage{
edgar_tickers(
  include_ticker_information = F,
  join_sic = T,
  snake_names = F,
  return_message = T
)
}
\arguments{
\item{include_ticker_information}{if \code{TRUE} returns ticker information}

\item{return_message}{}
}
\description{
SEC listed public companys
}
\examples{
edgar_tickers()
}
\concept{SEC EDGAR}
