% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adv_functions.R
\name{extract_fee_references}
\alias{extract_fee_references}
\title{Extract fee references}
\usage{
extract_fee_references(data, word_threshhold = 5, print_sentences = TRUE)
}
\arguments{
\item{data}{A \code{tibble} with the OCR'd brochur data}

\item{print_sentences}{if \code{TRUE} prints fee reference data}
}
\description{
Extract fee references
}
