% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finra.R
\name{finra_people}
\alias{finra_people}
\title{FINRA registered people}
\usage{
finra_people(search_name = NULL, ocr_pdf = TRUE, return_message = TRUE)
}
\arguments{
\item{search_name}{vector of names to search}

\item{ocr_pdf}{if \code{TRUE} returns OCR'd broker report PDF}

\item{return_message}{return a message upon parsing \code{TRUE, FALS}}

\item{score_threshold}{matching score threshold for the search name
if \code{NULL} there is no threshold}
}
\value{
a \code{data frame}
}
\description{
This function returns information for any
Financial Industry Regulatory Authority [FINRA] registered
entity.
}
\note{
Use \code{\link{finra_people()}} for registered people
}
\examples{
finra_people(search_name = 'Llyod Blankfein', ocr_pdf = TRUE)
}
\references{
\href{http://www.finra.org/}{FINRA}
}
\seealso{
Other FINRA: 
\code{\link{adv_managers_metadata}()},
\code{\link{finra_entities}()}

Other IAPD: 
\code{\link{adv_managers_current_period_summary}()},
\code{\link{adv_managers_filings}()},
\code{\link{adv_managers_metadata}()},
\code{\link{adv_managers_periods_summaries}()},
\code{\link{finra_entities}()}
}
\concept{FINRA}
\concept{IAPD}
\concept{person search}
