% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fred_api.R
\name{fred_terms_ids}
\alias{fred_terms_ids}
\title{FRED Terms Series IDs}
\usage{
fred_terms_ids(
  search_terms = NULL,
  use_json_api = TRUE,
  snake_names = F,
  return_message = TRUE
)
}
\arguments{
\item{search_terms}{vector of search terms}

\item{use_json_api}{if \code{TRUE} uses faster json API but results limted to 100
default \code{FALSE}}

\item{return_message}{if \code{TRUE} returns message}
}
\value{
a \code{tibble}
}
\description{
This function returns
matching FRED series ID for a given
search term.  When using
`use_json_api` search results limited to 100,
}
\examples{
fred_terms_ids(search_terms = c("China Debt", "China Housing"))
}
