% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trading_view.R
\name{generate_trade_view_metric_query}
\alias{generate_trade_view_metric_query}
\title{Generate tradeview metric query}
\usage{
generate_trade_view_metric_query(filter = data_frame(left =
  "market_cap_basic", operation = "nempty"), symbols = list(query =
  list(types = c("stock", "fund", "dr"))), metrics = c("change",
  "change_abs", "close", "description", "earnings_per_share_basic_ttm",
  "market_cap_basic", "name", "number_of_employees", "price_earnings_ttm",
  "sector", "volume"), sort = list(sortBy = "market_cap_basic", sortOrder
  = "desc"), options = list(lang = "eng"), range = c(0, 1500000))
}
\arguments{
\item{range}{}
}
\description{
Generate tradeview metric query
}
