% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adv_functions.R
\name{get_data_adv_managers_brochures}
\alias{get_data_adv_managers_brochures}
\title{OCR registerd managers ADV brochures}
\usage{
get_data_adv_managers_brochures(search_names = NULL, crd_ids = NULL,
  split_pages = TRUE)
}
\arguments{
\item{search_names}{names of the companies you want to search}

\item{crd_ids}{CRDs you want to search}

\item{split_pages}{do you want to split the brochure into multiple pages}
}
\description{
OCR registerd managers ADV brochures
}
\examples{
get_data_adv_managers_brochures(search_names = c('137 Ventures', 'Divco'), crd_ids = 156663, split_pages = TRUE)
}

