% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adv_functions.R
\name{get_data_adv_managers_current_period_summary}
\alias{get_data_adv_managers_current_period_summary}
\title{Get ADV filing data for the most recent filing period}
\usage{

  get_data_adv_managers_current_period_summary(file_directory = "Desktop/adv_data")
}
\arguments{
\item{file_directory}{}
}
\description{
Get ADV filing data for the most recent filing period
}

