% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adv_functions.R
\name{get_data_adv_managers_metadata}
\alias{get_data_adv_managers_metadata}
\title{Get ADV metadata for specified ADV or search name}
\usage{
get_data_adv_managers_metadata(search_names = "Divco", crd_ids = NULL,
  return_message = T)
}
\arguments{
\item{search_names}{Names of the entities you want to search}

\item{crd_ids}{CRD ids you want to search}

\item{return_message}{}
}
\description{
Get ADV metadata for specified ADV or search name
}
\examples{
get_data_adv_managers_metadata(search_names = c('Divco'))
}

