% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adv_functions.R
\name{get_data_adv_managers_periods_summaries}
\alias{get_data_adv_managers_periods_summaries}
\title{ADV managers periods data}
\usage{
get_data_adv_managers_periods_summaries(periods = c("2006-06"),
  all_periods = FALSE, only_most_recent = FALSE,
  include_exempt = TRUE, nest_data = FALSE, return_message = TRUE)
}
\arguments{
\item{periods}{dates in year-month form}

\item{all_periods}{include all periods}

\item{nest_data}{return a nested data frame}

\item{return_message}{return a message after parsing data}

\item{is_exempt}{exempt, non-exempt filers}
}
\value{
where \code{nest_data} is \code{TRUE} a nested data_frame by period and type of filer,
where \code{nest_data} is \code{FALSE} a data_frame
}
\description{
This function returns monthly summary
information for every ADV filing manager
from 2006 onwards.
}
\examples{
\dontrun{
get_data_adv_managers_periods_summaries(periods = c("2006-06", "2016-12", "2017-01"), all_periods = FALSE, is_exempt = c(FALSE,TRUE), only_most_recent = FALSE, nest_data = FALSE)

get_data_adv_managers_periods_summaries(only_most_recent = TRUE)
}
}
\seealso{
Other IAPD: \code{\link{get_data_adv_managers_current_period_summary}},
  \code{\link{get_data_adv_managers_filings}},
  \code{\link{get_data_adv_managers_metadata}},
  \code{\link{get_data_finra_entities}},
  \code{\link{get_data_finra_people}}

Other ADV: \code{\link{get_data_adv_managers_current_period_summary}},
  \code{\link{get_data_adv_managers_filings}}

Other entity search: \code{\link{get_data_adv_managers_current_period_summary}},
  \code{\link{get_data_adv_managers_filings}},
  \code{\link{get_data_adv_managers_metadata}},
  \code{\link{get_data_finra_entities}},
  \code{\link{get_data_nareit_entities}},
  \code{\link{get_data_reit_funds}},
  \code{\link{get_data_rf_leis}},
  \code{\link{get_data_rf_sec_13F_companies}},
  \code{\link{get_data_sec_bankruptcies}},
  \code{\link{get_data_sec_broker_dealers}},
  \code{\link{get_data_sec_ciks}},
  \code{\link{get_data_sec_closed_end_funds}},
  \code{\link{get_data_sec_cusips}},
  \code{\link{get_data_sec_filer}},
  \code{\link{get_data_sec_filing_entities}},
  \code{\link{get_data_sec_investment_companies}},
  \code{\link{get_data_sec_money_market_funds}},
  \code{\link{get_data_sec_municipal_advisors}},
  \code{\link{get_data_sec_xbrl_periods}},
  \code{\link{get_data_securities_offerings}},
  \code{\link{get_data_us_public_companies}},
  \code{\link{get_data_ycombinator_alumni}}

Other fund data: \code{\link{get_data_adv_managers_current_period_summary}},
  \code{\link{get_data_adv_managers_filings}},
  \code{\link{get_data_adv_managers_metadata}},
  \code{\link{get_data_reit_funds}},
  \code{\link{get_data_sec_ciks}},
  \code{\link{get_data_sec_closed_end_funds}},
  \code{\link{get_data_sec_money_market_funds}}
}
\concept{ADV}
\concept{IAPD}
\concept{entity search}
\concept{fund data}
