% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venture_data_functions.R
\name{get_data_cb_unicorns}
\alias{get_data_cb_unicorns}
\title{Get CB insight unicorn data}
\usage{
get_data_cb_unicorns(return_message = TRUE)
}
\arguments{
\item{return_message}{return a message after data parsing \code{TRUE, FALSE}}
}
\description{
Get CB insight unicorn data
}
\examples{
get_data_cb_unicorns(return_message = TRUE)
}

