% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtcc.R
\name{get_data_dtcc_most_recent_trades}
\alias{get_data_dtcc_most_recent_trades}
\title{Get DTCC most recent cleared trades}
\usage{
get_data_dtcc_most_recent_trades(assets = NULL, nest_data = TRUE,
  return_message = TRUE)
}
\arguments{
\item{assets}{type of DTCC cleared financial product \code{NULL, COMMODITIES, CREDITS, EQUITIES, FOREX, RATES}}

\item{return_message}{}
}
\description{
Get DTCC most recent cleared trades
}

