% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adv_functions.R
\name{get_data_finra_entities}
\alias{get_data_finra_entities}
\title{FINRA registered entities}
\usage{
get_data_finra_entities(entity_names = NULL, ocr_pdf = TRUE,
  score_threshold = 0.2, return_message = TRUE)
}
\arguments{
\item{entity_names}{vector of names to search}

\item{ocr_pdf}{if \code{TRUE} returns OCR'd broker report PDF}

\item{score_threshold}{matching score threshold for the search name
if \code{NULL} there is no threshold}

\item{return_message}{return a message upon parsing \code{TRUE, FALS}}
}
\value{
a \code{data frame}
}
\description{
This function returns information for any
Financial Industry Regulatory Authority [FINRA] registered
entity.
}
\note{
Use \code{\link{get_data_finra_people}} for registered people
}
\examples{
get_data_finra_entities(entity_names = c("EJF", "Blackstone", "Next Play", "Simple Capital"), score_threshold = .25,
ocr_pdf = TRUE)
}
\references{
\href{http://www.finra.org/}{FINRA}
}
\seealso{
Other FINRA: \code{\link{get_data_adv_managers_metadata}},
  \code{\link{get_data_finra_people}}

Other IAPD: \code{\link{get_data_adv_managers_current_period_summary}},
  \code{\link{get_data_adv_managers_filings}},
  \code{\link{get_data_adv_managers_metadata}},
  \code{\link{get_data_adv_managers_periods_summaries}},
  \code{\link{get_data_finra_people}}

Other entity search: \code{\link{get_data_adv_managers_current_period_summary}},
  \code{\link{get_data_adv_managers_filings}},
  \code{\link{get_data_adv_managers_metadata}},
  \code{\link{get_data_adv_managers_periods_summaries}},
  \code{\link{get_data_nareit_entities}},
  \code{\link{get_data_reit_funds}},
  \code{\link{get_data_rf_leis}},
  \code{\link{get_data_rf_sec_13F_companies}},
  \code{\link{get_data_sec_bankruptcies}},
  \code{\link{get_data_sec_broker_dealers}},
  \code{\link{get_data_sec_ciks}},
  \code{\link{get_data_sec_closed_end_funds}},
  \code{\link{get_data_sec_cusips}},
  \code{\link{get_data_sec_filer}},
  \code{\link{get_data_sec_filing_entities}},
  \code{\link{get_data_sec_investment_companies}},
  \code{\link{get_data_sec_money_market_funds}},
  \code{\link{get_data_sec_municipal_advisors}},
  \code{\link{get_data_sec_xbrl_periods}},
  \code{\link{get_data_securities_offerings}},
  \code{\link{get_data_us_public_companies}},
  \code{\link{get_data_ycombinator_alumni}}
}
\concept{FINRA}
\concept{IAPD}
\concept{entity search}
