% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adv_functions.R
\name{get_data_finra_people}
\alias{get_data_finra_people}
\title{Get FINRA registered people and associated metadata}
\usage{
get_data_finra_people(search_names = NULL, ocr_pdf = TRUE,
  return_message = TRUE)
}
\arguments{
\item{search_names}{vector of names to search}

\item{return_message}{return a message upon parsing \code{TRUE, FALS}}
}
\description{
Get FINRA registered people and associated metadata
}
\examples{
get_data_finra_people(search_name = 'Llyod Blankfein', ocr_pdf = TRUE)
}

