% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debt_functions.R
\name{get_data_index_symbol_current_value}
\alias{get_data_index_symbol_current_value}
\title{Get current specified index value}
\usage{
get_data_index_symbol_current_value(symbol = "^TNX", return_wide = F)
}
\arguments{
\item{return_wide}{}
}
\description{
Get current specified index value
}
\examples{
get_data_index_symbol_current_value("^TNX")
}

