% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgar_rf.R
\name{get_data_location_codes}
\alias{get_data_location_codes}
\title{Locations}
\usage{
get_data_location_codes()
}
\value{
a \code{data_frame}
}
\description{
This function returns location codes
and the name the location
}
\examples{
get_data_location_codes()
}
\seealso{
Other dictionary: \code{\link{get_data_sec_rules}},
  \code{\link{get_data_sic_naics_codes}},
  \code{\link{get_dictionary_all_fred_series_ids}},
  \code{\link{get_dictionary_sec_filing_codes}},
  \code{\link{get_dictionary_sec_form_codes}},
  \code{\link{get_dictionary_sic_codes}},
  \code{\link{get_reit_entity_dictionary}}
}
