% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debt_functions.R
\name{get_data_monthly_periods}
\alias{get_data_monthly_periods}
\title{Gets Monthly Period DF}
\usage{
get_data_monthly_periods(start_date = "2016-06-01", term_years = 25,
  term_months = 0)
}
\arguments{
\item{term_months}{}
}
\description{
Gets Monthly Period DF
}

