% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicies.R
\name{get_data_msci_indicies}
\alias{get_data_msci_indicies}
\title{MSCI indicies}
\usage{
get_data_msci_indicies()
}
\value{
\code{data_frame}
}
\description{
Returns all MSCI indicies.  This function
can be used to find indicies to search with
\code{\link{get_data_msci_indicies_constituents}}
to specify indicies to extract constituents.
}
\examples{
get_data_msci_indicies()
}
\seealso{
Other MSCI: \code{\link{get_data_msci_indicies_constituents}},
  \code{\link{get_data_msci_realtime_index_values}}
}
