% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicies.R
\name{get_data_msci_realtime_index_values}
\alias{get_data_msci_realtime_index_values}
\title{Get most recent MSCI index values}
\usage{
get_data_msci_realtime_index_values(return_wide = TRUE,
  return_message = TRUE)
}
\arguments{
\item{return_wide}{return a wide data frame \code{TRUE, FALSEs}}

\item{return_message}{return a message \code{TRUE, FALSE}}
}
\description{
Get most recent MSCI index values
}
\examples{
get_data_msci_realtime_index_values()
}

