% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nareit.R
\name{get_data_nareit_annual_subsector_returns}
\alias{get_data_nareit_annual_subsector_returns}
\title{NAREIT annual sub sector returns}
\usage{
get_data_nareit_annual_subsector_returns(return_wide = TRUE,
  return_message = TRUE)
}
\arguments{
\item{return_wide}{\code{TRUE} return data in wide form}

\item{return_message}{\code{TRUE} return a message after data import}
}
\value{
\code{data_frame}
}
\description{
This function returns annual subsector returns
for NAREIT sub sectors from 1994 to prior year-end.
}
\examples{
\dontrun{
get_data_nareit_annual_subsector_returns(return_wide = TRUE, return_message = TRUE)
}
}
\references{
\href{http://nareit.org}{National Association of Real Estate Investment Trusts}
}
\seealso{
Other NAREIT: \code{\link{get_data_nareit_capital_raises}},
  \code{\link{get_data_nareit_constituent_years}},
  \code{\link{get_data_nareit_entities}},
  \code{\link{get_data_nareit_industry_tracker}},
  \code{\link{get_data_nareit_mergers_acquisitions}},
  \code{\link{get_data_nareit_monthly_returns}},
  \code{\link{get_data_nareit_notable_properties}},
  \code{\link{get_data_nareit_property_msa}},
  \code{\link{get_data_nareit_state_info}},
  \code{\link{get_data_reit_funds}}

Other index values: \code{\link{get_data_crsp_indicies_returns}},
  \code{\link{get_data_fred_symbols}},
  \code{\link{get_data_msci_realtime_index_values}},
  \code{\link{get_data_nareit_monthly_returns}}
}
\concept{NAREIT}
\concept{index values}
