% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nareit.R
\name{get_data_nareit_constituent_years}
\alias{get_data_nareit_constituent_years}
\title{Get NAREIT index constuiency data by year}
\usage{
get_data_nareit_constituent_years(years = 1991:2016, resolve_names = TRUE,
  nest_data = TRUE, return_message = TRUE)
}
\arguments{
\item{years}{years of data you wa}

\item{return_message}{}
}
\description{
Get NAREIT index constuiency data by year
}
\examples{
get_data_nareit_constituent_years(years = 1991:2016, resolve_names = TRUE, nest_data = TRUE, return_messag = TRUE)
}

