% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nareit.R
\name{get_data_nareit_industry_tracker}
\alias{get_data_nareit_industry_tracker}
\title{Get NAREIT Sector Tracking Metrics by Quarter}
\usage{
get_data_nareit_industry_tracker(return_wide = TRUE, return_message = TRUE)
}
\description{
Get NAREIT Sector Tracking Metrics by Quarter
}
\examples{
get_data_nareit_industry_tracker(capital_type = NULL, nest_data = FALSE,return_message = TRUE)
}

