% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nareit.R
\name{get_data_nareit_mergers_acquisitions}
\alias{get_data_nareit_mergers_acquisitions}
\title{Get NAREIT merger & acquisition data by year}
\usage{
get_data_nareit_mergers_acquisitions(pages = 32:33, nest_data = FALSE,
  return_message = TRUE)
}
\arguments{
\item{pages}{location of the pages in the REIT monthly file, only use this if the location of the M&A table changes}

\item{nest_data}{return a nested data frame \code{TRUE, FALSE}}

\item{return_message}{return a message \code{TRUE, FALSE}}
}
\description{
Get NAREIT merger & acquisition data by year
}
\examples{
get_data_nareit_mergers_acquisitions(nest_data = FALSE, return_message = TRUE)
}

