% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nareit.R
\name{get_data_nareit_notable_properties}
\alias{get_data_nareit_notable_properties}
\title{Get NAREIT notable properties}
\usage{
get_data_nareit_notable_properties()
}
\description{
Get NAREIT notable properties
}
\examples{
get_data_nareit_notable_properties()
}

