% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nareit.R
\name{get_data_nareit_property_msa}
\alias{get_data_nareit_property_msa}
\title{Get NAREIT Member Property Holdings by MSA}
\usage{
get_data_nareit_property_msa(nest_data = TRUE, return_message = TRUE)
}
\arguments{
\item{return_message}{return a message \code{TRUE, FALSE}}
}
\description{
Get NAREIT Member Property Holdings by MSA
}
\examples{
get_data_nareit_property_msa(nest_data = TRUE, return_message = TRUE)
}

