% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sec_functions.R
\name{get_data_sec_foia_requests}
\alias{get_data_sec_foia_requests}
\title{Get SEC FOIA Requests}
\usage{
get_data_sec_foia_requests(search_years = 2006:2016, return_message = TRUE)
}
\arguments{
\item{return_message}{}
}
\description{
Get SEC FOIA Requests
}

