% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sec_functions.R
\name{get_data_sec_xbrl_periods}
\alias{get_data_sec_xbrl_periods}
\title{SEC quarterly XBRL data}
\usage{
get_data_sec_xbrl_periods(only_most_recent = TRUE, years = NULL,
  quarters = NULL, tables = NULL, only_all = TRUE,
  assign_to_environment = TRUE, return_message = TRUE)
}
\arguments{
\item{only_most_recent}{\code{TRUE} search only the most recent period}

\item{years}{years to include starting in 2009}

\item{quarters}{quarters to search \itemize{
\item\code{NULL}(default): all quarters,
\item \code{1}: Q1
\item \code{2}: Q2
\item \code{3}: Q3
\item \code{4}: Q4
}}

\item{tables}{underlying XBRL tables you want to include\itemize{
\item \code{NULL}: all tables (default)
\item \code{Presentation}: presentation data
\item \code{Values}: XBRL value data
\item \code{Filers}: filer data
\item \code{Taxonomy}: taxonomy data
}}

\item{only_all}{\code{TRUE} returns a joined file of the four tables}

\item{assign_to_environment}{\code{TRUE} assign the results into your environment}

\item{return_message}{\code{TRUE} return a message after data import}

\item{nest_data}{\code{TRUE} return nested data frame}
}
\value{
nested \code{data_frame} or \code{data_frame} if \code{nest_data = FALSE}
}
\description{
This function returns quarterly full \href{https://en.wikipedia.org/wiki/XBRL}{XBRL} dumps
for all SEC registered public companies begining in 2009.
}
\examples{
\dontrun{
get_data_sec_xbrl_periods(only_most_recent = TRUE, only_all = TRUE, assign_to_environment = TRUE, return_message = TRUE)
}

}
\references{
\href{http://sec.gov}{The Securities and Exchange Commission}
}
\seealso{
Other SEC: \code{\link{get_data_mmf_owned_debt_securities}},
  \code{\link{get_data_recent_insider_trades}},
  \code{\link{get_data_rf_leis}},
  \code{\link{get_data_sec_13F_companies}},
  \code{\link{get_data_sec_bankruptcies}},
  \code{\link{get_data_sec_broker_dealers}},
  \code{\link{get_data_sec_ciks}},
  \code{\link{get_data_sec_closed_end_funds}},
  \code{\link{get_data_sec_cusips}},
  \code{\link{get_data_sec_failed_to_deliver_securities}},
  \code{\link{get_data_sec_filer}},
  \code{\link{get_data_sec_filing_entities}},
  \code{\link{get_data_sec_foia_requests}},
  \code{\link{get_data_sec_investment_companies}},
  \code{\link{get_data_sec_money_market_funds}},
  \code{\link{get_data_sec_municipal_advisors}},
  \code{\link{get_data_sec_securities_filing_counts}},
  \code{\link{get_data_sec_securities_metrics_by_exchange}},
  \code{\link{get_data_securities_offerings}},
  \code{\link{get_data_us_public_companies}},
  \code{\link{get_dictionary_sec_filing_codes}},
  \code{\link{get_dictionary_sec_form_codes}}

Other entity search: \code{\link{get_data_adv_managers_current_period_summary}},
  \code{\link{get_data_adv_managers_filings}},
  \code{\link{get_data_adv_managers_metadata}},
  \code{\link{get_data_adv_managers_periods_summaries}},
  \code{\link{get_data_finra_entities}},
  \code{\link{get_data_nareit_entities}},
  \code{\link{get_data_reit_funds}},
  \code{\link{get_data_rf_leis}},
  \code{\link{get_data_sec_13F_companies}},
  \code{\link{get_data_sec_bankruptcies}},
  \code{\link{get_data_sec_broker_dealers}},
  \code{\link{get_data_sec_ciks}},
  \code{\link{get_data_sec_closed_end_funds}},
  \code{\link{get_data_sec_cusips}},
  \code{\link{get_data_sec_filer}},
  \code{\link{get_data_sec_filing_entities}},
  \code{\link{get_data_sec_investment_companies}},
  \code{\link{get_data_sec_money_market_funds}},
  \code{\link{get_data_sec_municipal_advisors}},
  \code{\link{get_data_securities_offerings}},
  \code{\link{get_data_us_public_companies}},
  \code{\link{get_data_ycombinator_alumni}}

Other public company data: \code{\link{get_data_sec_failed_to_deliver_securities}},
  \code{\link{get_data_sec_securities_metrics_by_exchange}}
}

