% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sec_functions.R
\name{get_data_sec_xbrl_periods}
\alias{get_data_sec_xbrl_periods}
\title{Get SEC quarterly XBRL data dumps}
\usage{
get_data_sec_xbrl_periods(only_most_recent = TRUE, years = NULL,
  quarters = NULL, tables = NULL, only_all = TRUE,
  assign_to_environment = TRUE, return_message = TRUE)
}
\arguments{
\item{return_message}{}
}
\description{
Get SEC quarterly XBRL data dumps
}

