% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_and_filed_functions.R
\name{get_data_sic_codes}
\alias{get_data_sic_codes}
\title{Get SIC code Classifications}
\usage{
get_data_sic_codes(filter_duplicates = TRUE)
}
\arguments{
\item{filter_duplicates}{}
}
\description{
Get SIC code Classifications
}

