% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicies.R
\name{get_data_sp500_constituents}
\alias{get_data_sp500_constituents}
\title{Get S&P 500 constituent data}
\usage{
get_data_sp500_constituents(return_message = TRUE)
}
\arguments{
\item{return_message}{return a message \code{TRUE, FALSE}}
}
\description{
Get S&P 500 constituent data
}
\examples{
get_data_sp500_constituents(return_message = TRUE)
}

