% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venture_data_functions.R
\name{get_data_ycombinator_alumni}
\alias{get_data_ycombinator_alumni}
\title{Y Combinator data}
\usage{
get_data_ycombinator_alumni(nest_data = FALSE, return_message = TRUE)
}
\arguments{
\item{nest_data}{\code{TRUE} return nested data frame}

\item{return_message}{\code{TRUE} return a message after data import
includes information about a random YC company}
}
\value{
where \code{nest_data} is \code{TRUE} a nested data_frame by batch,
where \code{nest_data} is \code{FALSE} a data_frame
}
\description{
This function returns information on
Y Combinator alumni and active particpants.
}
\examples{
get_data_ycombinator_alumni(nest_data = FALSE)
}
\references{
\href{https://www.ycombinator.com/}{YCombinator}
}
\seealso{
Other venture capital: \code{\link{calculate_valuation_post_money}},
  \code{\link{get_data_cb_unicorns}}

Other entity search: \code{\link{get_data_adv_managers_current_period_summary}},
  \code{\link{get_data_adv_managers_filings}},
  \code{\link{get_data_adv_managers_metadata}},
  \code{\link{get_data_adv_managers_periods_summaries}},
  \code{\link{get_data_finra_entities}},
  \code{\link{get_data_nareit_entities}},
  \code{\link{get_data_reit_funds}},
  \code{\link{get_data_rf_leis}},
  \code{\link{get_data_rf_sec_13F_companies}},
  \code{\link{get_data_sec_bankruptcies}},
  \code{\link{get_data_sec_broker_dealers}},
  \code{\link{get_data_sec_ciks}},
  \code{\link{get_data_sec_closed_end_funds}},
  \code{\link{get_data_sec_cusips}},
  \code{\link{get_data_sec_filer}},
  \code{\link{get_data_sec_filing_entities}},
  \code{\link{get_data_sec_investment_companies}},
  \code{\link{get_data_sec_money_market_funds}},
  \code{\link{get_data_sec_municipal_advisors}},
  \code{\link{get_data_securities_offerings}},
  \code{\link{get_data_us_public_companies}}
}
\concept{entity search}
\concept{venture capital}
