% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venture_data_functions.R
\name{get_data_ycombinator_alumni}
\alias{get_data_ycombinator_alumni}
\title{Get data on Y-Combinator alumni companies}
\usage{
get_data_ycombinator_alumni(return_message = T)
}
\arguments{
\item{return_message}{}
}
\description{
Get data on Y-Combinator alumni companies
}

