% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_and_filed_functions.R
\name{get_dictionary_form_d_categories}
\alias{get_dictionary_form_d_categories}
\title{Form-D dictionary}
\usage{
get_dictionary_form_d_categories()
}
\value{
a \code{data_frame}
}
\description{
This function returns searchable
industries for parsed SEC Form-D
filings
}
\examples{
get_dictionary_form_d_categories()
}

