% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_and_filed_functions.R
\name{get_dictionary_sec_filing_codes}
\alias{get_dictionary_sec_filing_codes}
\title{Get SEC filing code dictionary}
\usage{
get_dictionary_sec_filing_codes()
}
\description{
Get SEC filing code dictionary
}
\examples{
get_dictionary_sec_filing_codes()
}

