% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgar_rf.R
\name{get_dictionary_sec_rules}
\alias{get_dictionary_sec_rules}
\alias{get_dictionary_sec_rules}
\alias{get_dictionary_sec_rules}
\title{SEC Rule dictionary}
\usage{
get_dictionary_sec_rules()

get_dictionary_sec_rules()

get_dictionary_sec_rules()
}
\value{
a \code{data_frame}

a \code{data_frame}

a \code{data_frame}
}
\description{
This function retuns a
dictionary of SEC rules

This function retuns a
dictionary of SEC rules

This function retuns a
dictionary of SEC rules
}
\examples{
get_dictionary_sec_rules()
get_dictionary_sec_rules()
get_dictionary_sec_rules()
}
