% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_and_filed_functions.R
\name{get_dictionary_sec_rules}
\alias{get_dictionary_sec_rules}
\title{Get SEC rule dictionary}
\usage{
get_dictionary_sec_rules()
}
\description{
Get SEC rule dictionary
}
\examples{
get_dictionary_sec_rules()
}

