% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fred_api.R
\name{read_rda_file}
\alias{read_rda_file}
\title{read RDA file}
\usage{
read_rda_file(
 
    file = "https://github.com/abresler/FRED_Dictionaries/blob/master/data/fred_series_data.rda?raw=true",
  return_tibble = TRUE
)
}
\description{
read RDA file
}
