% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sec_2.R
\name{sic_filing_count}
\alias{sic_filing_count}
\title{SIC Counts}
\usage{
sic_filing_count(
  sic = NULL,
  join_sic = TRUE,
  snake_names = FALSE,
  use_all_sic_codes = FALSE,
  return_message = TRUE,
  unformat = FALSE
)
}
\arguments{
\item{sic}{vector of SIC codes}

\item{join_sic}{if \code{TRUE} joins SIC data}

\item{snake_names}{if \code{TRUE} returns snake case names}

\item{use_all_sic_codes}{uses all SIC codes}

\item{return_message}{if \code{TRUE} returns a message}

\item{unformat}{if \code{TRUE} returns unformatted values}
}
\value{
A tibble containing filing counts by SIC code
}
\description{
SIC Counts
}
\examples{
\dontrun{
sic_filing_count(sic = 6282)
}
}
