% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sec_2.R
\name{sic_filing_count}
\alias{sic_filing_count}
\title{SIC Counts}
\usage{
sic_filing_count(
  sic = NULL,
  join_sic = T,
  snake_names = F,
  use_all_sic_codes = F,
  return_message = T,
  unformat = F
)
}
\arguments{
\item{sic}{vector of SIC codes}

\item{use_all_sic_codes}{uses all SIC codes}

\item{return_message}{}

\item{join_sec}{if \code{TRUE} joins SIC data}
}
\description{
SIC Counts
}
