% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trading_view.R
\name{tv_tickers_news}
\alias{tv_tickers_news}
\title{Tradingview tickers news}
\usage{
tv_tickers_news(
  tickers = c("FB", "AAPL", "NFLX", "GOOG", "VNO", "EQR", "BXP"),
  return_message = TRUE,
  nest_data = FALSE
)
}
\value{

}
\description{
Returns news data for specified tickers
}
\examples{
tv_tickers_news(tickers = c("VNO", "AVB", "PEI"), return_message = TRUE, nest_data = FALSE)
}
